% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_wald.R
\name{ci_wald}
\alias{ci_wald}
\title{Confidence Interval based on Wald-test approximation}
\usage{
ci_wald(model, ci = 0.95, dof = Inf)
}
\arguments{
\item{model}{A statistical model.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{dof}{Degrees of Freedom.}
}
\description{
Confidence Interval based on Wald-test approximation
}
