% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_distribution.R
\docType{data}
\name{classify_distribution}
\alias{classify_distribution}
\title{Machine learning model trained to classify distributions}
\format{An object of class \code{train} (inherits from \code{train.formula}) of length 21.}
\usage{
classify_distribution
}
\description{
Mean accuracy and Kappa of 0.96.
}
\keyword{datasets}
