% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_f.R
\name{cohens_f}
\alias{cohens_f}
\title{Cohen's f}
\usage{
cohens_f(model)
}
\arguments{
\item{model}{an ANOVA object.}
}
\value{
Cohen's f values.
}
\description{
Computation of Cohen's f for ANOVAs.
}
\examples{
library(parameters)

df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
cohens_f(model)

model <- anova(lm(Sepal.Length ~ Sepal.Big * Species, data = df))
cohens_f(model)

}
