% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_paramerers.R
\name{compare_paramerers}
\alias{compare_paramerers}
\title{Compare model parameters of multiple models}
\usage{
compare_paramerers(...)
}
\arguments{
\item{...}{One or more regression model objects. Regression models may be
of different model types.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Compute and extract model parameters of multiple regression models.
  See \code{\link{model_parameters}} for further details.
}
