% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_d_to_r.R
\name{d_to_r}
\alias{d_to_r}
\alias{convert_d_to_r}
\title{Convert standardized difference d to correlation r.}
\usage{
d_to_r(d)

convert_d_to_r(d)
}
\arguments{
\item{d}{A standardized difference value (Cohen's d).}
}
\description{
Convert standardized difference (Cohen's d) to correlation r.
}
\examples{
d_to_r(d=1.1547)

}
\references{
\itemize{
  \item Borenstein, Michael, et al. "Converting among effect sizes." Introduction to meta-analysis (2009): 45-49.
}
}
