% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.factor_to_numeric}
\alias{.factor_to_numeric}
\title{Safe transformation from factor/character to numeric}
\usage{
.factor_to_numeric(x)
}
\description{
Safe transformation from factor/character to numeric
}
\keyword{internal}
