% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filter_component}
\alias{.filter_component}
\title{for models with zero-inflation component, return required component of model-summary}
\usage{
.filter_component(dat, component)
}
\description{
for models with zero-inflation component, return required component of model-summary
}
\keyword{internal}
