% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_bentler}
\alias{.n_factors_bentler}
\title{Bentler and Yuan's Procedure}
\usage{
.n_factors_bentler(x = NULL, cormatrix = NULL, nobs = NULL,
  eigen_values = NULL, type = "FA")
}
\description{
Bentler and Yuan's Procedure
}
\keyword{internal}
