% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_cng}
\alias{.n_factors_cng}
\title{Cattell-Nelson-Gorsuch CNG Indices}
\usage{
.n_factors_cng(x = NULL, cormatrix = NULL, nobs = NULL,
  eigen_values = NULL, type = "FA")
}
\description{
Cattell-Nelson-Gorsuch CNG Indices
}
\keyword{internal}
