% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_mreg}
\alias{.n_factors_mreg}
\title{Multiple Regression Procedure}
\usage{
.n_factors_mreg(eigen_values = NULL, model = "factors")
}
\description{
Multiple Regression Procedure
}
\keyword{internal}
