% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_scree}
\alias{.n_factors_scree}
\title{Non Graphical Cattel's Scree Test}
\usage{
.n_factors_scree(eigen_values = NULL, model = "factors")
}
\description{
Non Graphical Cattel's Scree Test
}
\keyword{internal}
