% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_sescree}
\alias{.n_factors_sescree}
\title{Standard Error Scree and Coefficient of Determination Procedures}
\usage{
.n_factors_sescree(x = NULL, cormatrix = NULL, nobs = NULL,
  eigen_values = NULL, type = "FA")
}
\description{
Standard Error Scree and Coefficient of Determination Procedures
}
\keyword{internal}
