% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_variances.R
\name{.random_sd_ci}
\alias{.random_sd_ci}
\title{(Not exporting; just using importFrom and document())}
\usage{
.random_sd_ci(
  model,
  out,
  ci_method,
  ci,
  corr_param,
  sigma_param,
  component = NULL
)
}
\value{

}
\description{
(Not exporting; just using importFrom and document())
}
