% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_to_zero}
\alias{.recode_to_zero}
\title{Recode a variable so its lowest value is beginning with zero}
\usage{
.recode_to_zero(x)
}
\description{
Recode a variable so its lowest value is beginning with zero
}
\keyword{internal}
