% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon_squared.R
\name{epsilon_squared}
\alias{epsilon_squared}
\title{Epsilon Squared.}
\usage{
epsilon_squared(model)
}
\arguments{
\item{model}{an ANOVA object.}
}
\value{
Cohen's f values.
}
\description{
Computation of Epsilon Squared for ANOVAs.
}
\examples{
library(parameters)

df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
epsilon_squared(model)

model <- anova(lm(Sepal.Length ~ Sepal.Big * Species, data = df))
epsilon_squared(model)

}
