% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_test.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equivalence_test}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{equivalence_test}}}
}}

