% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scales.R
\name{extract_scales}
\alias{extract_scales}
\title{Extract Scales from Principal Component Analysis (PCA)}
\usage{
extract_scales(x, n_items = NULL)
}
\arguments{
\item{x}{An object returned by \code{\link{principal_components}}.}

\item{n_items}{Number of required (i.e. non-missing) items to build the sum score. If \code{NULL}, the value is chosen to match half of the number of columns in a data frame.}
}
\value{
A data frame with subscales, which are mean sum scores for all items from each component.
}
\description{
To be done
}
\details{
\code{extrac_scales()} takes the results from \code{\link{principal_components}}
  and extracts the variables for each component found by the PCA. Then, for
  each of these "subscales", row means are calculated (which equals adding
  up the single items and dividing by the number of items). This results in
  a sum score for each component from the PCA, which is on the same scale as
  the original, single items that were used to compute the PCA.
}
\examples{
library(parameters)
pca <- principal_components(mtcars[, 1:7], n = 2, rotation = "varimax")

# PCA extracted two components
pca

# assignment of items to each component
component_columns(pca)

# now we want to have sum scores for each component
extract_scales(pca)

# compare to manually computed sum score for 2nd component, which
# consists of items "hp" and "qsec"
(mtcars$hp + mtcars$qsec) / 2
}
