% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_find_distribution.R
\name{find_distribution}
\alias{find_distribution}
\title{Classify the variable's distribution using machine learning}
\usage{
find_distribution(x, probabilities = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{probabilities}{Return a dataframe containing the probabilities of belonging to each distribution type.}
}
\description{
This function uses an internal random forest model to classify the variable's distribution.
}
\examples{
find_distribution(rnorm(100))
find_distribution(rpois(100, lambda = 4))
find_distribution(runif(100))
}
