% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_algorithm.R
\name{format_algorithm}
\alias{format_algorithm}
\title{Model Algorithm formatting}
\usage{
format_algorithm(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
Model Algorithm formatting
}
\examples{
model <- lm(Sepal.Length ~ Species, data = iris)
format_algorithm(model)
}
