% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Convert number to words}
\usage{
format_number(x, textual = TRUE, ...)
}
\arguments{
\item{x}{Number.}

\item{textual}{Return words. If \code{FALSE}, will run \code{\link{format_value}}.}

\item{...}{Arguments to be passed to \code{\link{format_value}} if \code{textual} is \code{FALSE}.}
}
\value{
A formatted string.
}
\description{
Convert number to words. The code has been adapted from here https://github.com/ateucher/useful_code/blob/master/R/numbers2words.r
}
\examples{
format_number(2)
format_number(45)
format_number(324.68765)
}
