% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_order.R
\name{format_order}
\alias{format_order}
\title{Order (first, second, ...) formatting}
\usage{
format_order(order, textual = TRUE, ...)
}
\arguments{
\item{order}{value or vector of orders.}

\item{textual}{Return number as words. If \code{FALSE}, will run \code{\link[insight]{format_value}}.}

\item{...}{Arguments to be passed to \code{\link[insight]{format_value}} if \code{textual} is \code{FALSE}.}
}
\value{
A formatted string.
}
\description{
Format order.
}
\examples{
format_order(2)
format_order(8)
format_order(25, textual = FALSE)
}
