% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p_adjust.R
\name{format_p_adjust}
\alias{format_p_adjust}
\title{Format the name of the p-value adjustment methods}
\usage{
format_p_adjust(method)
}
\arguments{
\item{method}{Name of the method.}
}
\value{
String.
}
\description{
Format the name of the p-value adjustment methods.
}
\examples{
library(parameters)

format_p_adjust("holm")
format_p_adjust("bonferroni")
}
