% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pd.R
\name{format_pd}
\alias{format_pd}
\title{Probability of direction (pd) formatting}
\usage{
format_pd(pd, stars = FALSE, stars_only = FALSE, name = "pd")
}
\arguments{
\item{pd}{Probability of direction (pd).}

\item{stars}{Add significance stars (e.g., p < .001***).}

\item{stars_only}{Return only significance stars.}

\item{name}{Name prefixing the text. Can be \code{NULL}.}
}
\value{
A formatted string.
}
\description{
Probability of direction (pd) formatting
}
\examples{
format_pd(0.12)
format_pd(c(0.12, 1, 0.9999, 0.98, 0.995, 0.96), name = NULL)
format_pd(c(0.12, 1, 0.9999, 0.98, 0.995, 0.96), stars = TRUE)
}
