% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rope.R
\name{format_rope}
\alias{format_rope}
\title{Percentage in ROPE formatting}
\usage{
format_rope(rope_percentage, name = "in ROPE")
}
\arguments{
\item{rope_percentage}{Value or vector of percentages in ROPE.}

\item{name}{Name prefixing the text. Can be \code{NULL}.}
}
\value{
A formatted string.
}
\description{
Percentage in ROPE formatting
}
\examples{
format_rope(c(0.02, 0.12, 0.357, 0))
format_rope(c(0.02, 0.12, 0.357, 0), name = NULL)
}
