% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Dataframe and Tables Pretty Formatting}
\usage{
format_table(x, sep = " | ")
}
\arguments{
\item{x}{A dataframe.}

\item{sep}{Column separator.}
}
\value{
A data.frame in character format.
}
\description{
Dataframe and Tables Pretty Formatting
}
\examples{
library(parameters)

cat(format_table(iris))

}
