% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMP_get_priors.R
\name{get_priors}
\alias{get_priors}
\title{Get information about priors specifications}
\usage{
get_priors(model, ...)
}
\arguments{
\item{model}{A Bayesian statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Get information about priors specifications
}
\examples{
\dontrun{
library(rstanarm)

model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width, data = iris)
get_priors(model)


library(brms)

model <- brms::brm(Sepal.Length ~ Petal.Width, data = iris)
get_priors(model)
}
}
