% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIP_marginal_parameters.R
\name{marginal_parameters}
\alias{marginal_parameters}
\alias{marginal_parameters.MixMod}
\title{Marginal Parameters for Mixed Models

Calculates marginal coefficients from fitted generalized linear mixed models.}
\usage{
marginal_parameters(model, ...)

\method{marginal_parameters}{MixMod}(model, std_errors = FALSE,
  link_fun = NULL, M = 3000, K = 100, seed = 1,
  cores = max(parallel::detectCores() - 1, 1), sandwich = FALSE, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}

\item{std_errors}{Logical indicating whether standard errors are to be computed.}

\item{link_fun}{A function transforming the mean of the repeated measurements outcome to the linear predictor scale. Typically, this derived from the family argument of mixed_model..}

\item{M}{Numeric scalar denoting the number of Monte Carlo samples.}

\item{K}{Numeric scalar denoting the number of samples from the sampling distribution of the maximum likelihood estimates.}

\item{seed}{Integer denoting the seed for the random number generation.}

\item{cores}{Integer giving the number of cores to use; applicable only when std_errors = TRUE.}

\item{sandwich}{Logical; if TRUE robust/sandwich standard errors are used in the calculations.}
}
\description{
Marginal Parameters for Mixed Models

Calculates marginal coefficients from fitted generalized linear mixed models.
}
\examples{
\dontrun{
model <- GLMMadaptive::mixed_model(vs ~ mpg,
  random = ~ 1 | cyl, data = mtcars,
  family = binomial()
)
marginal_parameters(model)
}

}
\references{
Hedeker, D., du Toit, S. H., Demirtas, H., & Gibbons, R. D. (2018). A note on marginalization of regression parameters from mixed models of binary outcomes. Biometrics, 74(1), 354-361.
}
\seealso{
\href{https://drizopoulos.github.io/GLMMadaptive/reference/marginal_coefs.html}{GLMMadaptive::marginal_coefs}.
}
