% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bootstrap.lm.R
\name{model_bootstrap.lm}
\alias{model_bootstrap.lm}
\title{Model bootstrapping}
\usage{
\method{model_bootstrap}{lm}(model, n = 1000, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{n}{The number of bootstrap replicates.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Model bootstrapping
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
model_bootstrap(model)
}
