% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bootstrap.R
\name{model_bootstrap}
\alias{model_bootstrap}
\title{Model bootstrapping}
\usage{
model_bootstrap(model, n = 1000, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{n}{The number of bootstrap replicates.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Bootstrap the model n times to return a data.frame of estimates.
}
