% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.gam.R
\name{model_parameters.gam}
\alias{model_parameters.gam}
\title{Parameters of Generalized Additive (Mixed) Models}
\usage{
\method{model_parameters}{gam}(model, ci = 0.95, bootstrap = FALSE,
  iterations = 1000, ...)
}
\arguments{
\item{model}{A gam/gamm model.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{bootstrap}{Should estimates be based on bootstrapped model? If \code{TRUE}, then arguments of \link[parameters:model_parameters.stanreg]{Bayesian regressions} apply (see also \code{\link[parameters:parameters_bootstrap]{parameters_bootstrap()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootstrapped frequentist models.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{\link[parameters:standardize.lm]{standardize()}}).}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Extract and compute indices and measures to describe parameters of generalized additive models (GAM(M)s).
}
\examples{
library(parameters)
library(mgcv)

dat <- gamSim(1, n = 400, dist = "normal", scale = 2)
model <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = dat)
model_parameters(model)
}
\seealso{
\code{\link[=standardize_names]{standardize_names()}} to rename
  columns into a consistent, standardized naming scheme.
}
