% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.glm.R
\name{model_parameters.glm}
\alias{model_parameters.glm}
\title{Linear Model Parameters}
\usage{
\method{model_parameters}{glm}(model, ci = 0.95, standardize = "refit",
  standardize_robust = FALSE, bootstrap = FALSE, iterations = 1000,
  ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \link{parameters_standardize}).}

\item{standardize_robust}{Robust standardization. See \link{parameters_standardize}.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{...}{Arguments passed to or from other methods (e.g., to \link[=standardize.lm]{standardize}).}
}
\description{
Parameters of general linear models.
}
\examples{
model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model, standardize = "refit")
}
