% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hclust.R, R/methods_kmeans.R
\name{model_parameters.hclust}
\alias{model_parameters.hclust}
\alias{model_parameters.kmeans}
\title{Parameters from Cluster Models (k-means, ...)}
\usage{
\method{model_parameters}{hclust}(model, verbose = TRUE, ...)

\method{model_parameters}{kmeans}(model, verbose = TRUE, ...)
}
\arguments{
\item{model}{Cluster model.}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Format cluster models obtained for example by \code{\link[=kmeans]{kmeans()}}.
}
\examples{
# hclust ---------------------------------
model <- hclust(dist(iris[1:4]))
rez <- model_parameters(model)
rez
library(parameters)

# K-means -------------------------------
model <- kmeans(iris[1:4], centers = 3)
rez <- model_parameters(model)
rez

# Clusters centers in long form
attributes(rez)$means
}
