% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.lavaan.R
\name{model_parameters.lavaan}
\alias{model_parameters.lavaan}
\title{Format CFA/SEM from the lavaan package}
\usage{
\method{model_parameters}{lavaan}(model, ci = 0.95,
  standardize = FALSE, type = c("regression", "correlation",
  "loading"), ...)
}
\arguments{
\item{model}{CFA or SEM created by the \code{lavaan::cfa} or \code{lavaan::sem} functions.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \code{\link{parameters_standardize}}).}

\item{type}{What type of links to return. Can be some of \code{c("regression", "correlation", "loading", "variance", "mean")}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Format CFA/SEM objects from the lavaan package (Rosseel, 2012).
}
\examples{
library(parameters)
\dontrun{
library(lavaan)

# Confirmatory Factor Analysis (CFA) ---------

structure <- ' visual  =~ x1 + x2 + x3
               textual =~ x4 + x5 + x6
               speed   =~ x7 + x8 + x9 '
model <- lavaan::cfa(structure, data=HolzingerSwineford1939)
model_parameters(model)

}

}
\references{
\itemize{
  \item Rosseel (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36.
}
}
