% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.Mclust.R
\name{model_parameters.Mclust}
\alias{model_parameters.Mclust}
\title{Parameters from Mixture Models}
\usage{
\method{model_parameters}{Mclust}(model, ...)
}
\arguments{
\item{model}{Mixture model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Format mixture models obtained for example by \code{mclust::Mclust}.
}
\examples{
library(parameters)
if (require("mclust")) {
  model <- mclust::Mclust(iris[1:4], verbose = FALSE)
  model_parameters(model)
}
}
