% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.pairwise.htest.R
\name{model_parameters.pairwise.htest}
\alias{model_parameters.pairwise.htest}
\title{Parameters from Pairwise tests}
\usage{
\method{model_parameters}{pairwise.htest}(model, ...)
}
\arguments{
\item{model}{Object of class \code{pairwise.htest}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters of pairwise h-tests (Wilcox tests, t-tests, etc.).
}
\examples{
smokers  <- c( 83, 90, 129, 70 )
patients <- c( 86, 93, 136, 82 )
p <- pairwise.prop.test(smokers, patients)
model_parameters(p)
}
