% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_PMCMRplus.R
\name{model_parameters.PMCMR}
\alias{model_parameters.PMCMR}
\title{Parameters from hypothesis tests from \code{PMCMRplus}}
\usage{
\method{model_parameters}{PMCMR}(model, ...)
}
\arguments{
\item{model}{Object of class \code{PMCMR} or\code{trendPMCMR} or \code{osrt}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters from hypothesis tests from \code{PMCMRplus}
}
\examples{
if (require("PMCMRplus")) {
  model <- kwAllPairsConoverTest(count ~ spray, data = InsectSprays)
  model_parameters(model)
}
}
