% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.bayesian.R
\name{model_parameters.stanreg}
\alias{model_parameters.stanreg}
\title{Bayesian Models Parameters}
\usage{
\method{model_parameters}{stanreg}(model, estimate = "median",
  dispersion = FALSE, ci = 0.9, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_full = TRUE, bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"), priors = TRUE, standardize = FALSE,
  standardize_robust = FALSE, iterations = 1000, ...)
}
\arguments{
\item{model}{Bayesian model.}

\item{estimate}{The \href{{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{point-estimate(s)}} to compute. Can be a character or a list with "median", "mean", "MAP" or "all".}

\item{dispersion}{if \code{TRUE}, computes indices of dispersion related to the estimate(s) (\code{SD} and \code{MAD} for \code{mean} and \code{median}, respectively).}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{ci_method}{The type of index used for Credible Interval. Can be \link{hdi} (default) or "quantile" (see \link{ci}).}

\item{test}{The \href{{https://easystats.github.io/bayestestR/articles/indicesEstimationComparison.html}{indices of effect existence}} to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".}

\item{rope_range}{\href{{https://easystats.github.io/bayestestR/rope}{ROPE's}} lower and higher bounds. Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}

\item{bf_prior}{Distribution representing a prior for the computation of \href{{https://easystats.github.io/bayestestR/bayesfactor_savagedickey}{Bayes factors}}. Used if the input is a posterior, otherwise (in the case of models) ignored.}

\item{diagnostic}{Diagnostic metrics to compute. Can be a character or a list with "ESS", "Rhat", "MCSE" or "all".}

\item{priors}{Add the prior used for each parameter.}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \link{standardize_parameters}).}

\item{standardize_robust}{Robust standardization. See \link{standardize_parameters}.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{...}{Arguments passed to or from other methods (e.g., to \link[=standardize.lm]{standardize}).}
}
\description{
Parameters of Bayesian models.
}
\examples{
\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Species, data=iris)
model_parameters(model, standardize = "full")

library(brms)
model <- brm(Sepal.Length ~ Species, data=iris)
model_parameters(model)
}
}
