% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_systemfit.R
\name{model_parameters.systemfit}
\alias{model_parameters.systemfit}
\title{Parameters from Linear Equation System Estimation Models}
\usage{
\method{model_parameters}{systemfit}(model, ci = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{model}{A model from \pkg{systemfit} package.}

\item{ci}{Confidence Interval (CI) level. Default to \code{0.95} (\verb{95\%}).}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to or from other methods. For instance, when
\code{bootstrap = TRUE}, arguments like \code{type} or \code{parallel} are
passed down to \code{bootstrap_model()}, and arguments like \code{ci_method}
are passed down to \code{\link[bayestestR:describe_posterior]{bayestestR::describe_posterior()}}.}
}
\description{
Parameters from Linear Equation System Estimation models (from
\pkg{systemfit} package).
}
