% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_wrs2.R
\name{model_parameters.yuen}
\alias{model_parameters.yuen}
\title{Parameters from \code{WRS2} objects}
\usage{
\method{model_parameters}{yuen}(model, verbose = TRUE, ...)
}
\arguments{
\item{model}{Object of class \code{yuen}.}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters from \code{WRS2} objects
}
\examples{
if (require("WRS2")) {
  model1 <- yuen(Anxiety ~ Group, data = spider)
  model_parameters(model1)

  before <- c(190, 210, 300, 240, 280, 170, 280, 250, 240, 220)
  after <- c(210, 210, 340, 190, 260, 180, 200, 220, 230, 200)
  model2 <- yuend(before, after)
  model_parameters(model2)
}
}
