% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_parameters.R
\name{n_parameters}
\alias{n_parameters}
\title{Count how many parameters in a model}
\usage{
n_parameters(x, ...)
}
\arguments{
\item{x}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Returns the number of parameters of a model.
}
