% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_to_d.R
\name{odds_to_d}
\alias{odds_to_d}
\title{Convert (log)odds to standardized difference.}
\usage{
odds_to_d(odds, log = FALSE)
}
\arguments{
\item{odds}{Odds values in vector or dataframe.}

\item{log}{Are these Log odds (such as in logistic models)?}
}
\description{
Convert (log)odds to standardized difference using Cohen's (1988) formula \eqn{\frac{\log(odds)\times\sqrt{3}}{\pi}}.
}
\examples{
odds_to_d(0.2)
odds_to_d(-1.45, log=TRUE)
}
