% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_odds_to_probs.R
\name{odds_to_probs}
\alias{odds_to_probs}
\alias{convert_odds_to_probs}
\title{Conversion between (log)odds and probabilies}
\usage{
odds_to_probs(odds, select = NULL, exclude = NULL, log = FALSE)

convert_odds_to_probs(odds, select = NULL, exclude = NULL,
  log = FALSE)
}
\arguments{
\item{odds}{Odds values in vector or dataframe.}

\item{select}{Character or list of of column names to be
transformed.}

\item{exclude}{Character or list of column names to be excluded
from transformation.}

\item{log}{Are these Log odds (such as in logistic models)?}
}
\description{
Enables a conversion between (log)odds and probabilities.
}
\examples{
odds_to_probs(-1.45)
}
