% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_squared.R
\name{omega_squared}
\alias{omega_squared}
\title{(Partial) Omega Squared.}
\usage{
omega_squared(model, partial = TRUE)
}
\arguments{
\item{model}{an ANOVA object.}

\item{partial}{Return partial omega squared.}
}
\description{
Computation of (Partial) Omega Squared for ANOVAs.
}
\examples{
library(parameters)

df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
omega_squared(model)

model <- anova(lm(Sepal.Length ~ Sepal.Big * Species, data = df))
omega_squared(model)

# model <- aov(Sepal.Length ~ Sepal.Big + Error(Species), data = df)
# omega_squared(model)


}
