% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_significance.R
\name{p_significance.lm}
\alias{p_significance.lm}
\title{Practical Significance (ps)}
\usage{
\method{p_significance}{lm}(x, threshold = "default", ci = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{x}{A statistical model.}

\item{threshold}{The threshold value that separates significant from
negligible effect, which can have following possible values:
\itemize{
\item \code{"default"}, in which case the range is set to \code{0.1} if input is a vector,
and based on \code{\link[bayestestR:rope_range]{rope_range()}} if a Bayesian model is provided.
\item a single numeric value (e.g., 0.1), which is used as range around zero
(i.e. the threshold range is set to -0.1 and 0.1)
\item a numeric vector of length two (e.g., \code{c(-0.2, 0.1)}).
}}

\item{ci}{Confidence Interval (CI) level. Default to \code{0.95} (\verb{95\%}).}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to other methods, e.g. \code{ci()}.}
}
\value{
A data frame.
}
\description{
Compute the probability of \strong{Practical Significance}
(\emph{ps}), which can be conceptualized as a unidirectional equivalence test.
It returns the probability that an effect is above a given threshold
corresponding to a negligible effect in the median's direction, considering
a parameter's confidence interval. In comparison the the \code{\link[=equivalence_test]{equivalence_test()}}
function, where the \emph{SGPV} (second generation p-value) describes the proportion
of the confidence interval that is \emph{inside} the ROPE, the value returned by
\code{p_significance()} describes the \emph{larger} proportion of the confidence
interval that is \emph{outside} the ROPE. This makes \code{p_significance()} comparable
to \code{\link[bayestestR:p_direction]{bayestestR::p_direction()}}, however, while \code{p_direction()} compares to
a point-null by default, \code{p_significance()} compares to a range-null.
}
\details{
\code{p_significance()} returns the proportion of the confidence interval
range (assuming a normally distributed, equal-tailed interval) that is
outside a certain range (the negligible effect, or ROPE, see argument
\code{threshold}). If there are values of the distribution both below and above
the ROPE, \code{p_significance()} returns the higher probability of a value being
outside the ROPE. Typically, this value should be larger than 0.5 to indicate
practical significance. However, if the range of the negligible effect is
rather large compared to the range of the confidence interval,
\code{p_significance()} will be less than 0.5, which indicates no clear practical
significance.
}
\note{
There is also a \href{https://easystats.github.io/see/articles/bayestestR.html}{\code{plot()}-method}
implemented in the \href{https://easystats.github.io/see/}{\strong{see}-package}.
}
\examples{
\dontshow{if (requireNamespace("bayestestR") && packageVersion("bayestestR") > "0.14.0") (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(qol_cancer)
model <- lm(QoL ~ time + age + education, data = qol_cancer)

p_significance(model)
p_significance(model, threshold = c(-0.5, 1.5))

# plot method
if (require("see", quietly = TRUE)) {
  result <- p_significance(model)
  plot(result)
}
\dontshow{\}) # examplesIf}
}
\seealso{
For more details, see \code{\link[bayestestR:p_significance]{bayestestR::p_significance()}}. See also
\code{\link[=equivalence_test]{equivalence_test()}}.
}
