% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_mfx.R
\name{p_value.poissonmfx}
\alias{p_value.poissonmfx}
\alias{p_value.betaor}
\alias{p_value.betamfx}
\title{p-values for Models with Zero-Inflation}
\usage{
\method{p_value}{poissonmfx}(model, component = c("all", "conditional", "marginal"), ...)

\method{p_value}{betaor}(model, component = c("all", "conditional", "precision"), ...)

\method{p_value}{betamfx}(
  model,
  component = c("all", "conditional", "precision", "marginal"),
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{component}{Should all parameters, parameters for the conditional model, precision-component or marginal effects be returned? \code{component} may be one of \code{"conditional"}, \code{"precision"}, \code{"marginal"} or \code{"all"} (default).}

\item{...}{Arguments passed down to \code{standard_error_robust()} when confidence intervals or p-values based on robust standard errors should be computed. Only available for models where \code{method = "robust"} is supported.}
}
\value{
The p-values.
}
\description{
This function attempts to return, or compute, p-values of hurdle and zero-inflated models.
}
\examples{
if (require("mfx")) {
  set.seed(12345)
  n <- 1000
  x <- rnorm(n)
  y <- rnegbin(n, mu = exp(1 + 0.5 * x), theta = 0.5)
  d <- data.frame(y, x)
  model <- poissonmfx(y ~ x, data = d)

  p_value(model)
  p_value(model, component = "marginal")
}
}
