% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_kenward.R
\name{p_value_kenward}
\alias{p_value_kenward}
\alias{dof_kenward}
\title{Compute p values using Kenward-Roger approximation}
\usage{
p_value_kenward(model, dof = NULL)

dof_kenward(model)
}
\arguments{
\item{model}{A statistical model.}

\item{dof}{Degrees of Freedom.}
}
\description{
The Wald-test approximation treats t-values as Wald z.
}
\examples{
\dontrun{
model <- lme4::lmer(Petal.Length ~ Sepal.Length + (1|Species), data=iris)
p_value_kenward(model)
model <- circus::merMod_1
p_value_kenward(model)
}
}
