% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof_ml1.R, R/p_value_ml1.R, R/se_ml1.R
\name{dof_ml1}
\alias{dof_ml1}
\alias{p_value_ml1}
\alias{se_ml1}
\title{p-values using m-l-1 heuristic}
\usage{
dof_ml1(model)

p_value_ml1(model, dof = NULL)

se_ml1(model)
}
\arguments{
\item{model}{A statistical model.}

\item{dof}{Degrees of Freedom.}
}
\value{
The p-values.
}
\description{
An approximate F-test based on a m-l-1 as suggested by Giesbrecht and Burns (1985).
}
\details{
\code{dof_ml1()} and \code{se_ml1()} are small helper-functions
to calculate approximated degrees of freedom and standard errors of model
parameters, based on the m-l-1 heuristic as suggested by Giesbrecht and
Burns.
}
\examples{
\donttest{
library(lme4)
model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
p_value_ml1(model)
}

}
\references{
Giesbrecht, F.G. and Burns, J.C. (1985), "Two-Stage Analysis Based on a Mixed Model: Large-sample Asymptotic Theory and Small-Sample Simulation Results", Biometrics 41: 853-862.
}
