% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_estimation.R
\name{p_value_robust}
\alias{p_value_robust}
\title{Robust p values. Superseded by the \verb{vcov*} arguments in \code{p_value()}}
\usage{
p_value_robust(
  model,
  vcov_estimation = "HC",
  vcov_args = NULL,
  component = "conditional",
  method = NULL,
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{vcov_estimation}{String, indicating the suffix of the
\verb{vcov*()}-function from the \pkg{sandwich} or \pkg{clubSandwich}
package, e.g. \code{vcov_estimation = "CL"} (which calls
\code{\link[sandwich:vcovCL]{sandwich::vcovCL()}} to compute clustered covariance matrix
estimators), or \code{vcov_estimation = "HC"} (which calls
\code{\link[sandwich:vcovHC]{sandwich::vcovHC()}} to compute
heteroskedasticity-consistent covariance matrix estimators).}

\item{vcov_args}{List of named vectors, used as additional arguments that are
passed down to the \pkg{sandwich}-function specified in
\code{vcov_estimation}.}

\item{component}{Model component for which parameters should be shown. See
the documentation for your object's class in \code{\link[=model_parameters]{model_parameters()}} for
further details.}

\item{method}{string For linear mixed models, \code{method} can be
\code{\link[=p_value_kenward]{"kenward"}} or
\code{\link[=p_value_satterthwaite]{"satterthwaite"}}.}

\item{...}{Additional arguments}
}
\description{
Robust p values. Superseded by the \verb{vcov*} arguments in \code{p_value()}
}
