% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof_satterthwaite.R, R/p_value_satterthwaite.R
\name{dof_satterthwaite}
\alias{dof_satterthwaite}
\alias{p_value_satterthwaite}
\title{p-values using Satterthwaite approximation}
\usage{
dof_satterthwaite(model)

p_value_satterthwaite(model, dof = NULL)
}
\arguments{
\item{model}{A statistical model.}

\item{dof}{Degrees of Freedom.}
}
\value{
The p-values.
}
\description{
An approximate F-test based on a Satterthwaite approximation as suggested by Giesbrecht \& Burns (1985).
}
\details{
\code{dof_satterthwaite()} and \code{se_satterthwaite()} are small helper-functions
to calculate approximated degrees of freedom and standard errors for model
parameters, based on the Giesbrecht \& Burns approach.
}
\examples{
\donttest{
library(lme4)
model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
p_value_satterthwaite(model)
}

}
\references{
Giesbrecht, F.G. and Burns, J.C. (1985), "Two-Stage Analysis Based on a Mixed Model: Large-sample Asymptotic Theory and Small-Sample Simulation Results", Biometrics 41: 853-862.
}
