% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_wald.R
\name{p_value_wald}
\alias{p_value_wald}
\alias{p_value_wald.merMod}
\title{Wald-test approximation for CIs and p-values}
\usage{
p_value_wald(model, ...)

\method{p_value_wald}{merMod}(model, dof = Inf, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed down to \code{standard_error_robust()} when confidence
intervals or p-values based on robust standard errors should be computed.}
}
\value{
A data frame.
}
\description{
The Wald-test approximation treats t-values as Wald z. Since the
t distribution converges to the z distribution as degrees of freedom
increase, this is like assuming infinite degrees of freedom. While this is
unambiguously anti-conservative, this approximation appears as reasonable
for reasonable sample sizes (Barr et al., 2013). That is, if we take the
p-value to measure the probability of a false positive, this approximation
produces a higher false positive rate than the nominal \verb{5\%} at \code{p = 0.05}.
}
\examples{
\donttest{
if (require("lme4")) {
  model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
  p_value_wald(model)
}
}
}
\references{
Barr, D. J. (2013). Random effects structure for testing interactions in
linear mixed-effects models. Frontiers in psychology, 4, 328.
}
