% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{class}
\name{parameter-class}
\alias{parameter}
\alias{parameter-class}
\alias{show,parameter-method}
\title{Bounded parameter objects}
\usage{
\S4method{show}{parameter}(object)
}
\arguments{
\item{object}{herp}
}
\description{
An S4 class to represent a bounded parameter for a model
}
\details{
No constructor function exists for \code{parameter} objects. Parameter
objects should be instantiated with the \code{new()} function directly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print details about the object to the console
}}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of length one, which acts as the parameters value}

\item{\code{name}}{A character vector of length one, which acts as a semantic label
for what the parameter value represents}

\item{\code{upper_bound}}{A numeric vector of length one. The \code{value} slot may
not be assinged a value greater than the value of the \code{upper_bound} slot.}

\item{\code{lower_bound}}{A numeric vector of length one. The \code{value} slot may
not be assinged a value less than the value of the \code{lower_bound} slot.}
}}
\examples{
new("parameter", value = .5, name = "binomial_p",
    upper_bound = 1, lower_bound = 0)
}

