% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_bootstrap.R
\name{parameters_bootstrap}
\alias{parameters_bootstrap}
\title{Parameters bootstrapping}
\usage{
parameters_bootstrap(model, iterations = 1000, centrality = "median",
  ci = 0.95, ci_method = "quantile", test = "p-value", ...)
}
\arguments{
\item{model}{Statistical model.}

\item{iterations}{The number of bootstrap replicates.}

\item{centrality}{The point-estimates (centrality indices) to compute. Can be a character or a list with "median", "mean", "MAP" or "all".}

\item{ci}{Value or vector of probability of the interval (between 0 and 1)
to be estimated. Named Credible Interval (CI) for consistency.}

\item{ci_method}{The type of index used for Credible Interval. Can be \link{hdi} (default) or "quantile" (see \link{ci}).}

\item{test}{The indices of effect existence to compute. Can be a character or a list with "p_direction", "rope", "p_map" or "bayesfactor".}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute bootstrapped parameters and their related indices such as Confidence Intervals (CI) and p-values.
}
\examples{
library(parameters)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
parameters_bootstrap(model)

}
\references{
Davison, A. C., & Hinkley, D. V. (1997). Bootstrap methods and their application (Vol. 1). Cambridge university press.
}
