% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_grouplevel.R
\name{parameters_groupspecific}
\alias{parameters_groupspecific}
\title{Extract group-specific parameters of mixed models}
\usage{
parameters_groupspecific(model, ...)
}
\arguments{
\item{model}{A mixed model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Extract parameters of each individual group in the context of mixed models.
}
\examples{
library(parameters)
if (require("lme4")) {
  data <- lme4::sleepstudy

  # Random intercept
  model <- lmer(Reaction ~ Days + (1 | Subject), data = data)
  parameters_groupspecific(model)

  # Random slope and intercept
  model <- lmer(Reaction ~ Days + (1 + Days  | Subject), data = data)
  parameters_groupspecific(model)

  # Nested random factors
  set.seed(12345)
  data$grp <- sample(1:5, size = 180, replace = TRUE)
  data$subgrp <- NA
  for (i in 1:5) {
    filter_group <- data$grp == i
    data$subgrp[filter_group] <- sample(1:30, size = sum(filter_group), replace = TRUE)
  }
  model <- lmer(Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject), data = data)
  parameters_groupspecific(model)

}
}
