% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_selection.R
\name{parameters_selection}
\alias{parameters_selection}
\alias{parameters_selection.lm}
\alias{parameters_selection.merMod}
\alias{parameters_selection.stanreg}
\title{Parameters Selection}
\usage{
parameters_selection(model, ...)

\method{parameters_selection}{lm}(model, ...)

\method{parameters_selection}{merMod}(model, ...)

\method{parameters_selection}{stanreg}(model, method = NULL,
  cross_validation = FALSE, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}

\item{method}{The method used in the variable selection. Can be NULL (default), "forward" or "L1". See \code{projpred::varsel}.}

\item{cross_validation}{Select with cross-validation.}
}
\description{
Automated selection of the 'best' parameters. For frequentist simple GLMs, it performes an AIC-based stepwise selection. For Bayesian models, it uses the \code{projpred} package.
}
\examples{
model <- lm(mpg ~ ., data = mtcars)
parameters_selection(model)

model <- lm(mpg ~ cyl * disp * hp * wt, data = mtcars)
parameters_selection(model)
\dontrun{
# lme4 -------------------------------------------
library(lme4)
model <- lmer(Sepal.Width ~ Sepal.Length * Petal.Width * Petal.Length + (1 | Species), data = iris)
parameters_selection(model)

# rstanarm -------------------------------------------
library(rstanarm)
model <- stan_glm(mpg ~ ., data = mtcars)
parameters_selection(model, cross_validation = TRUE)

model <- stan_glm(mpg ~ cyl * disp * hp, data = mtcars)
parameters_selection(model, cross_validation = FALSE)
}

}
