% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_table.R
\name{parameters_table}
\alias{parameters_table}
\title{Parameter table formatting}
\usage{
parameters_table(
  x,
  pretty_names = TRUE,
  stars = FALSE,
  digits = 2,
  ci_width = "auto",
  ci_brackets = TRUE,
  ci_digits = 2,
  p_digits = 3,
  ...
)
}
\arguments{
\item{x}{A data frame of model's parameters, as returned by \code{model_parameters()}. May also be a result from \code{broom::tidy()}.}

\item{pretty_names}{Return "pretty" (i.e. more human readable) parameter names.}

\item{stars}{Add significance stars (e.g., p < .001***).}

\item{digits}{Number of decimal places for numeric values (except confidence intervals and p-values).}

\item{ci_width}{Minimum width of the returned string for confidence intervals. If not \code{NULL} and width is larger than the string's length, leading whitespaces are added to the string. If \code{width="auto"}, width will be set to the length of the longest string.}

\item{ci_brackets}{Logical, if \code{TRUE} (default), CI-values are encompassed in square brackets (else in parantheses).}

\item{ci_digits}{Number of decimal places for confidence intervals.}

\item{p_digits}{Number of decimal places for p-values. May also be \code{"scientific"} for scientific notation of p-values.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame.
}
\description{
This functions takes a data frame with model parameters as input
  and formats certain columns into a more readable layout (like collapsing
  separate columns for lower and upper confidence interval values). Furthermore,
  column names are formatted as well.
}
\examples{
x <- model_parameters(lm(Sepal.Length ~ Species * Sepal.Width, data = iris))
as.data.frame(parameters_table(x))
as.data.frame(parameters_table(x, p_digits = "scientific"))
\donttest{
if (require("rstanarm")) {
  model <- stan_glm(Sepal.Length ~ Species, data = iris, refresh = 0, seed = 123)
  x <- model_parameters(model, ci = c(0.69, 0.89, 0.95))
  as.data.frame(parameters_table(x))
}}
}
