% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_kmeans.R
\name{predict.parameters_clusters}
\alias{predict.parameters_clusters}
\title{Predict method for parameters_clusters objects}
\usage{
\method{predict}{parameters_clusters}(object, newdata = NULL, names = NULL, ...)
}
\arguments{
\item{object}{a model object for which prediction is desired.}

\item{newdata}{data.frame}

\item{names}{character vector or list}

\item{...}{additional arguments affecting the predictions produced.}
}
\description{
Predict method for parameters_clusters objects
}
