% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.parameters_model.R
\name{print}
\alias{print}
\alias{print.parameters_model}
\title{Print model parameters}
\usage{
\method{print}{parameters_model}(
  x,
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  show_sigma = FALSE,
  ...
)
}
\arguments{
\item{x}{An object returned by \code{\link[=model_parameters]{model_parameters()}}.}

\item{pretty_names}{Return "pretty" (i.e. more human readable) parameter names.}

\item{split_components}{Logical, if \code{TRUE} (default), For models with
multiple components (zero-inflation, smooth terms, ...), each component is
printed in a separate table. If \code{FALSE}, model parameters are printed
in a single table and a \code{Component} column is added to the output.}

\item{select}{Character vector (or numeric index) of column names that should
be printed. If \code{NULL} (default), all columns are printed. The shortcut
\code{select = "minimal"} prints coefficient, confidence intervals and p-values,
while \code{select = "short"} prints coefficient, standard errors and p-values.}

\item{digits}{Number of decimal places for numeric values (except confidence intervals and p-values).}

\item{ci_digits}{Number of decimal places for confidence intervals.}

\item{p_digits}{Number of decimal places for p-values. May also be \code{"scientific"} for scientific notation of p-values.}

\item{show_sigma}{Logical, if \code{FALSE}, adds information about the residual
standard deviation.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
\code{NULL}
}
\description{
A \code{print()}-method for objects from \code{\link[=model_parameters]{model_parameters()}}.
}
\section{Notation of Interaction Terms}{

The notation of interaction terms in \code{format_parameters()}, and hence
also in \code{print()} for \code{model_parameters()}, uses both \code{*}
and \code{:}, depending on the type of interaction. Whenever
\emph{main effects} for all parameters of the interaction terms are present,
the interaction is denoted as \code{*}. When at least one of the interaction
terms is not included as main effect, the interaction is denoted as \code{:}.

\preformatted{m1 <- lm(mpg ~ qsec : wt + wt / drat, data = mtcars)
model_parameters(m1)
> Parameter   | Coefficient |   SE |          95\% CI | t(28) |      p
> -------------------------------------------------------------------
> (Intercept) |       35.89 | 2.34 | [ 31.09, 40.68] | 15.34 | < .001
> wt          |      -10.47 | 1.96 | [-14.48, -6.46] | -5.35 | < .001
> qsec : wt   |        0.28 | 0.09 | [  0.10,  0.46] |  3.17 | 0.004
> wt : drat   |        0.17 | 0.43 | [ -0.71,  1.05] |  0.39 | 0.701

m2 <- lm(mpg ~ qsec * wt, data = mtcars)
model_parameters(m2)
> Parameter   | Coefficient |    SE |          95\% CI | t(28) |     p
> -------------------------------------------------------------------
> (Intercept) |      -13.11 | 29.38 | [-73.29, 47.06] | -0.45 | 0.659
> qsec        |        2.76 |  1.63 | [ -0.58,  6.09] |  1.69 | 0.102
> wt          |        5.50 |  9.29 | [-13.53, 24.53] |  0.59 | 0.559
> qsec * wt   |       -0.59 |  0.52 | [ -1.65,  0.47] | -1.14 | 0.265}
}

\section{Interpretation of Interaction Terms}{

Note that the \emph{interpretation} of model coefficients depends on many
characteristics of the model. The number of parameters, and overall
performance of the model, can differ \emph{or not} between \code{a * b}
and \code{a / b}, suggesting that sometimes interaction terms give different
parameterizations of the same model, but other times it gives completely
different models (depending on \code{a} or \code{b} being factors of
covariates, included as main effects or not, etc.). Their interpretation
depends of the full context of the model, which should not be inferred
from the parameters table alone - rather, we recommend to use packages
to calculate estimated marginal means or marginal effects, such as
\CRANpkg{modelbased}, \CRANpkg{emmeans} or \CRANpkg{ggeffects}.
}

\examples{
\donttest{
library(parameters)
if (require("glmmTMB")) {
  model <- glmmTMB(
    count ~ spp + mined + (1 | site),
    ziformula = ~mined,
    family = poisson(),
    data = Salamanders
  )
  mp <- model_parameters(model)

  print(mp, pretty_names = FALSE)

  print(mp, split_components = FALSE)

  print(mp, select = c("Parameter", "Coefficient", "SE"))

  print(mp, select = "minimal")
}
}
}
