% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{qol_cancer}
\alias{qol_cancer}
\title{Sample data set}
\format{
A data frame with 564 rows and 7 variables:
\describe{
  \item{ID}{Patient ID}
  \item{QoL}{Quality of Life Score}
  \item{time}{Timepoint of measurement}
  \item{age}{Age in years}
  \item{phq4}{Patients' Health Questionnaire, 4-item version}
  \item{hospital}{Hospital ID, where patient was treated}
  \item{education}{Patients' educational level}
}
}
\description{
A sample data set with longitudinal data, used in the vignette describing the \code{demean()} function. Health-related quality of life from cancer-patients was measured at three time points (pre-surgery, 6 and 12 months after surgery).
}
\keyword{data}
