% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_parameters.R
\name{random_parameters}
\alias{random_parameters}
\title{Summary information from random effects}
\usage{
random_parameters(model)
}
\arguments{
\item{model}{A mixed effects model.}
}
\value{
A data frame with random effects statistics, including number of levels per
  random effect group, as well as complete observations in the model.
}
\description{
This function extracts the different variance components of a
  mixed model as well as further information related to random effects (ICC, R2)
  and returns the result as a data frame.
}
\details{
The variance components are obtained from \code{\link[insight]{get_variance}}
  and are denoted as following:
  \subsection{Within-subject (or residual) variance}{
    The residual variance, \ifelse{html}{\out{&sigma;<sup>2</sup><sub>&epsilon;</sub>}}{\eqn{\sigma^2_\epsilon}},
    is the sum of the distribution-specific variance and the variance due to additive dispersion.
    It indicates the \emph{within-subject variance}.
  }
  \subsection{Between-subject (or random intercept) variance}{
    The random intercept variance, or \emph{between-subject} variance
    (\ifelse{html}{\out{&tau;<sub>00</sub>}}{\eqn{\tau_{00}}}),
    is obtained from \code{VarCorr()}. It indicates how much groups
    or subjects differ from each other.
  }
  \subsection{Random slope variance}{
    The random slope variance (\ifelse{html}{\out{&tau;<sub>11</sub>}}{\eqn{\tau_{11}}})
    is obtained from \code{VarCorr()}. This measure is only available
    for mixed models with random slopes.
  }
  \subsection{Random slope-intercept correlation}{
    The random slope-intercept correlation
    (\ifelse{html}{\out{&rho;<sub>01</sub>}}{\eqn{\rho_{01}}})
    is obtained from \code{VarCorr()}. This measure is only available
    for mixed models with random intercepts and slopes.
  }
 For details regarding the calculation of R2 and ICC, see \code{\link[performance]{r2_nakagawa}}
 and \code{\link[performance]{icc}}.
}
\examples{
if (require("lme4")) {
  data(sleepstudy)
  model <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
  random_parameters(model)
}
}
