\name{ratios} 
\docType{data}
\alias{ratios}
\title{ratios: some decontextualized gene expression ratios,
  a biologically useless placeholder for data in your package.}

\description{
These 134 floating point numbers are of forgotten provenance, but useful
to show how to store, retrieve and document data sets in an R package.}

\examples{
  data (ratios)
  print (fivenum (ratios))

}

\keyword{datasets}
