% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_test.R, R/methods_bayestestR.R,
%   R/n_parameters.R, R/p_direction.R, R/p_significance.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equivalence_test}
\alias{ci}
\alias{n_parameters}
\alias{p_direction}
\alias{p_significance}
\alias{standardize_names}
\alias{supported_models}
\alias{print_html}
\alias{print_md}
\alias{display}
\alias{describe_distribution}
\alias{demean}
\alias{rescale_weights}
\alias{visualisation_recipe}
\alias{kurtosis}
\alias{skewness}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR:ci]{ci()}}, \code{\link[bayestestR:equivalence_test]{equivalence_test()}}, \code{\link[bayestestR:p_direction]{p_direction()}}, \code{\link[bayestestR:p_significance]{p_significance()}}}

  \item{datawizard}{\code{\link[datawizard:demean]{demean()}}, \code{\link[datawizard:describe_distribution]{describe_distribution()}}, \code{\link[datawizard:kurtosis]{kurtosis()}}, \code{\link[datawizard:rescale_weights]{rescale_weights()}}, \code{\link[datawizard:skewness]{skewness()}}, \code{\link[datawizard:visualisation_recipe]{visualisation_recipe()}}}

  \item{insight}{\code{\link[insight:display]{display()}}, \code{\link[insight:n_parameters]{n_parameters()}}, \code{\link[insight:print_html]{print_html()}}, \code{\link[insight:print_md]{print_md()}}, \code{\link[insight:standardize_names]{standardize_names()}}, \code{\link[insight:supported_models]{supported_models()}}}
}}

