% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{smoothness}
\alias{smoothness}
\title{Quantify the smoothness of a vector}
\usage{
smoothness(x, method = "cor")
}
\arguments{
\item{x}{Numeric vector (similar to a time series).}

\item{method}{Can be "diff" (the standard deviation of the standardized differences) or "cor" (default, lag-one autocorrelation).}
}
\description{
Quantify the smoothness of a vector
}
\examples{
smoothness(exp(-10:10), method="cor")
smoothness(exp(-10:10), method="diff")

}
\references{
https://stats.stackexchange.com/questions/24607/how-to-measure-smoothness-of-a-time-series-in-r
}
