% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_parameters.R
\name{sort_parameters}
\alias{sort_parameters}
\alias{sort_parameters.default}
\title{Sort parameters by coefficient values}
\usage{
sort_parameters(x, ...)

\method{sort_parameters}{default}(x, sort = "none", style = "easystats", ...)
}
\arguments{
\item{x}{A data frame or a \code{parameters_model} object.}

\item{...}{Arguments passed to or from other methods.}

\item{sort}{If \code{"none"} (default) do not sort, \code{"ascending"} sort by
increasing coefficient value, or \code{"descending"} sort by decreasing
coefficient value.}

\item{style}{Naming conventions the given object follows. This can either be
based on the naming conventions from the
\href{https://easystats.github.io/easystats/}{easystats-project}, or on
\pkg{broom}'s naming conventions.}
}
\value{
A sorted data frame or original object.
}
\description{
Sort parameters by coefficient values
}
\details{
The column containing model parameter estimates is called \code{Coefficient} in
\emph{easystats} convention and \code{estimate} in \emph{broom} convention. Therefore, if
you enter a data frame with a different naming convention for column
containing the said estimates, it will produce an error.
}
\examples{
# creating object to sort (can also be a regular data frame)
mod <- model_parameters(stats::lm(wt ~ am * cyl, data = mtcars))

# original output
mod

# sorted outputs
sort_parameters(mod, sort = "ascending")
sort_parameters(mod, sort = "descending")

}
