% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R, R/standardize.data.frame.R
\name{normalize}
\alias{normalize}
\alias{normalize.numeric}
\alias{normalize.factor}
\alias{normalize.grouped_df}
\alias{normalize.data.frame}
\alias{standardize.data.frame}
\title{Normalization}
\usage{
normalize(x, ...)

\method{normalize}{numeric}(x, ...)

\method{normalize}{factor}(x, ...)

\method{normalize}{grouped_df}(x, select = NULL, exclude = NULL, ...)

\method{normalize}{data.frame}(x, select = NULL, exclude = NULL, ...)

\method{standardize}{data.frame}(x, robust = FALSE, select = NULL,
  exclude = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}

\item{select}{For a data.frame, character or list of characters of column names to be
standardized.}

\item{exclude}{For a data.frame, character or list of characters of column names to be excluded from standardization.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}
}
\description{
Performs a normalization of data. This scales all numeric variables in the range 0 - 1.

Standardize (scale and reduce, Z-score) the data so that the values are expressed in terms of standard deviation (i.e., mean = 0, SD = 1) or Median Absolute Deviance (\code{robust = TRUE}; median = 0, MAD = 1). A \code{normalization} scales all numeric variables in the 0 - 1 range.
}
\examples{
summary(standardize(iris))
}
