% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.data.frame.R
\name{standardize.factor}
\alias{standardize.factor}
\title{Standardization}
\usage{
\method{standardize}{factor}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Performs a standardization of data or parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=standardize.data.frame]{Dataframes}}
 \item{\link[=standardize.lm]{Regression models}}
 }
}
\details{
Methods:
\itemize{
 \item \strong{mean}: \code{mean} and \code{sd} (default).
 \item \strong{median}: \code{median} and \code{mad}.
 \item \strong{2sd}: \code{mean} and two times the \code{sd}.
}
}
