% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.models.R
\name{standardize.lm}
\alias{standardize.lm}
\title{Model Standardization}
\usage{
\method{standardize}{lm}(x, robust = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Refit the model after standardizing the data.
}
\examples{
model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
coef(standardize(model))
}
