% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{standardize_names}
\alias{standardize_names}
\alias{standardize_names.parameters_model}
\title{Standardize column names}
\usage{
standardize_names(data, ...)

\method{standardize_names}{parameters_model}(data, style = c("easystats",
  "broom"), ...)
}
\arguments{
\item{data}{A data frame. Currently, only objects from \code{\link[=model_parameters]{model_parameters()}} are accepted.}

\item{style}{Standardization can either be based on the naming conventions from the easystats project, or on \pkg{broom}'s naming scheme.}
}
\value{
A data frame, with standardized column names.
}
\description{
Standardize column names from data frames, in particular objects returned from
\code{\link[=model_parameters]{model_parameters()}}, so column names are
consistent and the same for any model object.
}
\examples{
library(parameters)
model <- lm(mpg ~ wt + cyl, data = mtcars)
mp <- model_parameters(model)

as.data.frame(mp)
standardize_names(mp)
standardize_names(mp, style = "broom")
}
