% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_posteriors.R
\name{summarise_posteriors}
\alias{summarise_posteriors}
\title{Summarise posterior distributions.}
\usage{
summarise_posteriors(posteriors, ci = 0.9, estimate = "median",
  test = c("pd", "rope"), rope_bounds = "default", rope_full = TRUE)
}
\arguments{
\item{posteriors}{A dataframe of posterior draws.}

\item{ci}{Credible Interval (CI) level. Default to 0.90 (90\%).}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{test}{What \href{https://easystats.github.io/bayestestR/articles/3_IndicesExistenceComparison.html}{indices of effect existence} to compute. Can be a character or a list with "p_direction", "rope" or "p_map".}

\item{rope_bounds}{\href{https://easystats.github.io/bayestestR/articles/1_IndicesDescription.html#rope}{ROPE's} lower and higher bounds. Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}
}
\description{
Summarise posterior distributions.
}
