% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_markdown.R, R/to_table.R
\name{table_to_markdown.parameters_model}
\alias{table_to_markdown.parameters_model}
\alias{to_table.parameters_model}
\alias{to_table.parameters_sem}
\alias{to_table.parameters_stan}
\alias{to_table.parameters_efa_summary}
\alias{to_table.parameters_efa}
\alias{to_table.equivalence_test_lm}
\title{Export tables into different output formats}
\usage{
\method{table_to_markdown}{parameters_model}(
  x,
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ...
)

\method{to_table}{parameters_model}(
  x,
  format = "markdown",
  pretty_names = TRUE,
  split_components = TRUE,
  select = NULL,
  digits = 2,
  ci_digits = 2,
  p_digits = 3,
  ...
)

\method{to_table}{parameters_sem}(x, format = "markdown", digits = 2, ci_digits = 2, p_digits = 3, ...)

\method{to_table}{parameters_stan}(x, split_components = TRUE, select = NULL, ...)

\method{to_table}{parameters_efa_summary}(x, format = "markdown", digits = 3, ...)

\method{to_table}{parameters_efa}(
  x,
  format = "markdown",
  digits = 2,
  sort = FALSE,
  threshold = NULL,
  labels = NULL,
  ...
)

\method{to_table}{equivalence_test_lm}(x, format = "markdown", digits = 2, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=model_parameters]{model_parameters()}},
\code{\link[=simulate_parameters]{simulate_parameters()}},
\code{\link[=equivalence_test.lm]{equivalence_test()}} or
\code{\link[=principal_components]{principal_components()}}.}

\item{pretty_names}{Return "pretty" (i.e. more human readable) parameter names.}

\item{split_components}{Logical, if \code{TRUE} (default), For models with
multiple components (zero-inflation, smooth terms, ...), each component is
printed in a separate table. If \code{FALSE}, model parameters are printed
in a single table and a \code{Component} column is added to the output.}

\item{select}{Character vector (or numeric index) of column names that should
be printed. If \code{NULL} (default), all columns are printed. The shortcut
\code{select = "minimal"} prints coefficient, confidence intervals and p-values,
while \code{select = "short"} prints coefficient, standard errors and p-values.}

\item{digits}{Number of decimal places for numeric values (except confidence intervals and p-values).}

\item{ci_digits}{Number of decimal places for confidence intervals.}

\item{p_digits}{Number of decimal places for p-values. May also be \code{"scientific"} for scientific notation of p-values.}

\item{...}{Arguments passed to or from other methods.}

\item{format}{String, indicating the output format. Currently, only
\code{"markdown"} is supported.}
}
\value{
A character vector. If \code{format = "markdown"}, the return value
  will be a character vector in markdown-table format.
}
\description{
Export tables (i.e. data frame) into different output formats.
  \code{table_to_markdown()} is a alias for \code{to_table(format = "markdown")}.
}
\details{
\code{to_table()} is useful when the table-output from functions,
  which is usually printed as formatted text-table to console, should
  be formatted for pretty table-rendering in markdown documents, or if
  knitted from rmarkdown to PDF or Word files. See
  \href{https://easystats.github.io/parameters/articles/model_parameters_formatting.html}{vignette}
  for examples.
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
mp <- model_parameters(model)
to_table(mp)
}
