% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bootstrap.R
\name{as.data.frame.lm}
\alias{as.data.frame.lm}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{lm}(x, row.names = NULL, optional = FALSE,
  n = 1000, silent = FALSE, ...)
}
\arguments{
\item{x}{Any R object.}

\item{row.names}{Not used.}

\item{optional}{Not used.}

\item{n}{The number of bootstrap replicates.}

\item{silent}{Hide possible refit messages.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\description{
Coerce to a Data Frame
}
