% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_model.R
\name{bootstrap_model}
\alias{bootstrap_model}
\title{Model bootstrapping}
\usage{
bootstrap_model(model, iterations = 1000, verbose = FALSE, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{iterations}{The number of draws to simulate/bootstrap.}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of bootstrapped estimates.
}
\description{
Bootstrap a statistical model n times to return a data frame of estimates.
}
\section{Using with \code{emmeans}}{

The output can be passed directly to the various functions from the
\code{emmeans} package, to obtain bootstrapped estimates, contrasts, simple
slopes, etc, and their confidence intervals. These can then be passed to
\code{model_parameter()} to obtain standard errors, p-values, etc (see
example).
}

\examples{
\donttest{
if (require("boot")) {
  model <- lm(mpg ~ wt + factor(cyl), data = mtcars)
  b <- bootstrap_model(model)
  print(head(b))

  if (require("emmeans")) {
    est <- emmeans(b, consec ~ cyl)
    print(model_parameters(est))
  }
}
}
}
\seealso{
\code{\link{bootstrap_parameters}}, \code{\link{simulate_model}}, \code{\link{simulate_parameters}}
}
