% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ci.R, R/methods_glmmTMB.R
\name{ci.default}
\alias{ci.default}
\alias{ci.glmmTMB}
\title{Confidence Intervals (CI)}
\usage{
\method{ci}{default}(x, ci = 0.95, dof = NULL, method = NULL, robust = FALSE, ...)

\method{ci}{glmmTMB}(
  x,
  ci = 0.95,
  dof = NULL,
  method = "wald",
  robust = FALSE,
  component = "all",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A statistical model.}

\item{ci}{Confidence Interval (CI) level. Default to \code{0.95} (\verb{95\%}).}

\item{dof}{Number of degrees of freedom to be used when calculating
confidence intervals. If \code{NULL} (default), the degrees of freedom are
retrieved by calling \code{\link[=degrees_of_freedom]{degrees_of_freedom()}} with
approximation method defined in \code{method}. If not \code{NULL}, use this argument
to override the default degrees of freedom used to compute confidence
intervals.}

\item{method}{Method for computing degrees of freedom for confidence
intervals (CI) or p-values. May be one of \code{"analytical"}, \code{"any"}, \code{"ml1"},
\code{"betwithin"}, \code{"satterthwaite"}, \code{"kenward"}, \code{"wald"}, \code{"profile"},
\code{"boot"}, \code{"uniroot"}, \code{"residual"}, \code{"normal"}, or \code{"likelihood"}.
If \code{bootstrap = TRUE}, may be one of \code{"hdi"}, \code{"quantile"}, \code{"ci"}, \code{"eti"},
\code{"si"}, \code{"bci"}, or \code{"bcai"}. See argument \code{ci_method} in
\code{\link[bayestestR:describe_posterior]{bayestestR::describe_posterior()}}.}

\item{robust}{Logical, if \code{TRUE}, computes confidence intervals based on
robust standard errors. See \code{\link[=standard_error_robust]{standard_error_robust()}}.}

\item{...}{Arguments passed down to \code{\link[=standard_error_robust]{standard_error_robust()}}
when confidence intervals or p-values based on robust standard errors
should be computed.}

\item{component}{TODO.}
}
\value{
A data frame containing the CI bounds.
}
\description{
Compute confidence intervals (CI) for frequentist models.
}
\note{
\code{ci_robust()} resp. \code{ci(robust = TRUE)} rely on the \pkg{sandwich}
or \pkg{clubSandwich} package (the latter if \code{vcov_estimation = "CR"} for
cluster-robust standard errors) and will thus only work for those models
supported by those packages.
}
\examples{
\donttest{
library(parameters)
if (require("glmmTMB")) {
  model <- glmmTMB(
    count ~ spp + mined + (1 | site),
    ziformula = ~mined,
    family = poisson(),
    data = Salamanders
  )

  ci(model)
  ci(model, component = "zi")
}
}
}
