% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.merMod}
\alias{ci.merMod}
\title{Confidence Interval}
\usage{
\method{ci}{merMod}(model, ci = 0.95, method = c("wald", "boot"), ...)
}
\arguments{
\item{method}{For mixed models, can be \link[=ci_wald]{"wald"} (default) or "boot" (see \code{lme4::confint.merMod}).}
}
\description{
Confidence Interval
}
