% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.merMod}
\alias{ci.merMod}
\alias{ci.glm}
\title{Confidence Interval (CI)}
\usage{
\method{ci}{merMod}(x, ci = 0.95, method = c("wald", "boot"), ...)

\method{ci}{glm}(x, ci = 0.95, ...)
}
\arguments{
\item{x}{A statistical model.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{method}{For mixed models, can be \link[=ci_wald]{"wald"} (default) or "boot" (see \code{lme4::confint.merMod}).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data.frame containing the CI bounds.
}
\description{
Compute confidence intervals (CI) for frequentist models.
}
