% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci.merMod}
\alias{ci.merMod}
\title{Confidence Intervals}
\usage{
\method{ci}{merMod}(x, ci = 0.95, method = c("wald", "boot"), ...)
}
\arguments{
\item{x}{A statistical model.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{method}{For mixed models, can be \link[=ci_wald]{"wald"} (default) or "boot" (see \code{lme4::confint.merMod}).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compte confidence intervals.
}
