% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ci}
\alias{ci.glm}
\alias{ci.lm}
\title{Objects exported from other packages}
\usage{
\method{ci}{glm}(x, ci = 0.95, ...)

\method{ci}{lm}(x, ci = 0.95, ...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}}
}}

