% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_performance.R
\name{cluster_performance}
\alias{cluster_performance}
\alias{cluster_performance.kmeans}
\title{Performance of clustering models}
\usage{
cluster_performance(model, ...)

\method{cluster_performance}{kmeans}(model, ...)
}
\arguments{
\item{model}{Cluster model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute performance indices for clustering solutions.
}
\examples{
# kmeans
model <- kmeans(iris[1:4], 3)
cluster_performance(model)
}
