% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_paramerers.R
\name{compare_parameters}
\alias{compare_parameters}
\alias{compare_models}
\title{Compare model parameters of multiple models}
\usage{
compare_parameters(
  ...,
  ci = 0.95,
  effects = "fixed",
  component = "conditional",
  standardize = NULL,
  exponentiate = FALSE,
  df_method = NULL,
  p_adjust = NULL,
  verbose = TRUE,
  style = NULL
)

compare_models(
  ...,
  ci = 0.95,
  effects = "fixed",
  component = "conditional",
  standardize = NULL,
  exponentiate = FALSE,
  df_method = NULL,
  p_adjust = NULL,
  verbose = TRUE,
  style = NULL
)
}
\arguments{
\item{...}{One or more regression model objects, or objects returned by
\code{model_parameters()}. Regression models may be of different model
types.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{effects}{Should parameters for fixed effects (\code{"fixed"}), random effects (\code{"random"}) or both (\code{"all"}) be returned? Only applies to mixed models. May be abbreviated.}

\item{component}{Model component for which parameters should be shown. See
documentation for related model class in \code{\link{model_parameters}}.}

\item{standardize}{The method used for standardizing the parameters. Can be
\code{"refit"}, \code{"posthoc"}, \code{"smart"}, \code{"basic"},
\code{"pseudo"} or \code{NULL} (default) for no standardization. See
'Details' in \code{\link[effectsize]{standardize_parameters}}. Note that
robust estimation (i.e. \code{robust=TRUE}) of standardized parameters only
works when \code{standardize="refit"}.}

\item{exponentiate}{Logical, indicating whether or not to exponentiate the
the coefficients (and related confidence intervals). This is typical for,
say, logistic regressions, or more generally speaking: for models with log
or logit link. \strong{Note:} standard errors are also transformed (by
multiplying the standard errors with the exponentiated coefficients), to
mimic behaviour of other software packages, such as Stata.}

\item{df_method}{Method for computing degrees of freedom for confidence
intervals (CI). Only applies to models of class \code{glm} or \code{polr}.
May be \code{"profile"} or \code{"wald"}.}

\item{p_adjust}{Character vector, if not \code{NULL}, indicates the method to
adjust p-values. See \code{\link[stats]{p.adjust}} for details.}

\item{verbose}{Toggle warnings and messages.}

\item{style}{String, indicating which style of output is requested. Following
templates are possible:
\itemize{
  \item \code{"ci_p"}: Estimate, confidence intervals and asterisks for p-values.
  \item \code{"se_p"}: Estimate, standard errors and asterisks for p-values.
  \item \code{"ci"}: Estimate and confidence intervals, no asterisks for p-values.
  \item \code{"se"}: Estimate and standard errors, no asterisks for p-values.
  \item \code{"ci_p2"}: Estimate, confidence intervals and numeric p-values, in two columns.
  \item \code{"se_p2"}: Estimate, standard errors and numeric p-values, in two columns.
}}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Compute and extract model parameters of multiple regression models.
  See \code{\link{model_parameters}} for further details.
}
\note{
This function is in an early stage and does not yet cope with more
  complex models, and probably does not yet properly render all model
  components.
}
\examples{
data(iris)
lm1 <- lm(Sepal.Length ~ Species, data = iris)
lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
compare_parameters(lm1, lm2, lm3)
}
