% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_bartlett}
\alias{.n_factors_bartlett}
\title{Bartlett, Anderson and Lawley Procedures}
\usage{
.n_factors_bartlett(x = NULL, cormatrix = NULL, nobs = NULL,
  eigen_values = NULL, type = "FA")
}
\description{
Bartlett, Anderson and Lawley Procedures
}
\keyword{internal}
