% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_bentler}
\alias{.n_factors_bentler}
\title{Bentler and Yuan's Procedure}
\usage{
.n_factors_bentler(eigen_values = NULL, model = "factors",
  nobs = NULL)
}
\description{
Bentler and Yuan's Procedure
}
\keyword{internal}
