% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{.n_factors_mreg}
\alias{.n_factors_mreg}
\title{Multiple Regression Procedure}
\usage{
.n_factors_mreg(x = NULL, cormatrix = NULL, nobs = NULL,
  eigen_values = NULL, type = "FA")
}
\description{
Multiple Regression Procedure
}
\keyword{internal}
