% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_squared.R
\name{eta_squared}
\alias{eta_squared}
\title{(Partial) Omega Squared.}
\usage{
eta_squared(model, partial = TRUE, ci = NULL)
}
\arguments{
\item{model}{an ANOVA object.}

\item{partial}{Return partial eta squared.}

\item{ci}{Scalar between 0 and 1. If not \code{NULL}, lower and upper
confidence intervals are returned as well.}
}
\value{
Eta squared values.
}
\description{
Computation of (Partial) Eta Squared for ANOVAs.
}
\examples{
library(parameters)

df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
eta_squared(model)

model <- anova(lm(Sepal.Length ~ Sepal.Big * Species, data = df))
eta_squared(model)

\donttest{
# Don't work for now
model <- aov(Sepal.Length ~ Sepal.Big + Error(Species), data = df)
eta_squared(model)
}

}
