% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Dataframe and Tables Pretty Printing}
\usage{
format_table(x, sep = " | ")
}
\arguments{
\item{x}{A dataframe.}

\item{sep}{Column separator.}
}
\description{
Dataframe and Tables Pretty Printing
}
\examples{
library(parameters)

x <- model_parameters(lm(Sepal.Length ~ Species * Sepal.Width, data = iris))
cat(format_table(x))

}
