% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bootstrap.R
\name{model_bootstrap}
\alias{model_bootstrap}
\title{Model bootstrapping}
\usage{
model_bootstrap(model, n = 1000, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{n}{The number of bootstrap replicates.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
See the documentation for your object's class:
\itemize{
 \item{\link[=model_bootstrap.lm]{lm}}
 }
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
