% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bootstrap.R
\name{model_bootstrap}
\alias{model_bootstrap}
\title{Model bootstrapping}
\usage{
model_bootstrap(model, iterations = 1000, verbose = FALSE, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{iterations}{The number of bootstrap replicates.}

\item{verbose}{Hide possible refit messages.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data.frame.
}
\description{
Bootstrap a statistical model n times to return a data.frame of estimates.
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
head(model_bootstrap(model))
}
\seealso{
\code{\link{parameters_bootstrap}}, \code{\link{model_simulate}}, \code{\link{parameters_simulate}}
}
