% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.aov.R
\name{model_parameters.aov}
\alias{model_parameters.aov}
\alias{model_parameters.anova}
\alias{model_parameters.aovlist}
\title{ANOVAs Parameters}
\usage{
\method{model_parameters}{aov}(model, omega_squared = "partial", ...)

\method{model_parameters}{anova}(model, omega_squared = "partial", ...)

\method{model_parameters}{aovlist}(model, omega_squared = "partial", ...)
}
\arguments{
\item{model}{Object of class \link{aov}, \link{anova} or \code{aovlist}.}

\item{omega_squared}{Compute omega squared as indices of effect size. Can be \code{NULL}, "partial" (default) or "raw" for non-partial indices.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Parameters of ANOVAs.
}
\examples{
\dontrun{
df <- iris
df$Sepal.Big <- ifelse(df$Sepal.Width >= 3, "Yes", "No")

model <- aov(Sepal.Length ~ Sepal.Big, data = df)
model_parameters(model)

model <- anova(lm(Sepal.Length ~ Sepal.Big, data = df))
model_parameters(model)

model <- aov(Sepal.Length ~ Sepal.Big + Error(Species), data = df)
model_parameters(model)

model <- anova(lme4::lmer(Sepal.Length ~ Sepal.Big + (1 | Species), data = df))
model_parameters(model)
}
}
