% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.BFBayesFactor.R
\name{model_parameters.BFBayesFactor}
\alias{model_parameters.BFBayesFactor}
\title{BayesFactor objects Parameters}
\usage{
\method{model_parameters}{BFBayesFactor}(model, ci = 0.9,
  iterations = 4000, ...)
}
\arguments{
\item{model}{Object of class \code{BFBayesFactor}.}

\item{ci}{Credible Interval (CI) level. Default to 0.90 (90\%).}

\item{iterations}{Number of posterior draws.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Parameters of BayesFactor objects.
}
