% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.glm.R
\name{model_parameters.glm}
\alias{model_parameters.glm}
\title{Linear Model Parameters}
\usage{
\method{model_parameters}{glm}(model, ci = 0.95, standardize = FALSE,
  bootstrap = FALSE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{standardize}).}
}
\description{
Parameters of general linear models.
}
\examples{
model <- glm(vs ~ wt + cyl, data = mtcars, family = "binomial")
model_parameters(model, standardize = TRUE)
}
