% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.htest.R,
%   R/model_parameters.pairwise.htest.R
\name{model_parameters.htest}
\alias{model_parameters.htest}
\alias{model_parameters.pairwise.htest}
\title{Parameters from hypothesis tests}
\usage{
\method{model_parameters}{htest}(
  model,
  cramers_v = NULL,
  phi = NULL,
  ci = 0.95,
  bootstrap = FALSE,
  ...
)

\method{model_parameters}{pairwise.htest}(model, ...)
}
\arguments{
\item{model}{Object of class \code{htest} or \code{pairwise.htest}.}

\item{cramers_v, phi}{Compute Cramer's V or phi as index of effect size.
Can be \code{"raw"} or \code{"adjusted"} (effect size will be bias-corrected).
Only applies to objects from \code{chisq.test()}.}

\item{ci}{Level of confidence intervals for Cramer's V or phi. Currently only
applies to objects from \code{chisq.test()}.}

\item{bootstrap}{Should estimates be bootstrapped?}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters of h-tests (correlations, t-tests, chi-squared, ...).
}
\examples{
model <- cor.test(mtcars$mpg, mtcars$cyl, method = "pearson")
model_parameters(model)

model <- t.test(iris$Sepal.Width, iris$Sepal.Length)
model_parameters(model)

model <- t.test(mtcars$mpg ~ mtcars$vs)
model_parameters(model)

model <- t.test(iris$Sepal.Width, mu = 1)
model_parameters(model)

data(airquality)
airquality$Month <- factor(airquality$Month, labels = month.abb[5:9])
model <- pairwise.t.test(airquality$Ozone, airquality$Month)
model_parameters(model)

smokers <- c(83, 90, 129, 70)
patients <- c(86, 93, 136, 82)
model <- pairwise.prop.test(smokers, patients)
model_parameters(model)
}
