% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.kmeans.R
\name{model_parameters.kmeans}
\alias{model_parameters.kmeans}
\title{Parameters from Cluster Models (k-means, ...)}
\usage{
\method{model_parameters}{kmeans}(model, ...)
}
\arguments{
\item{model}{Cluster model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Format cluster models obtained for example by \code{\link{kmeans}}.
}
\examples{
library(parameters)

model <- kmeans(iris[1:4], centers = 3)
model_parameters(model)
}
