% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.lavaan.R
\name{model_parameters.lavaan}
\alias{model_parameters.lavaan}
\title{Format CFA/SEM from the (b)lavaan package}
\usage{
\method{model_parameters}{lavaan}(
  model,
  ci = 0.95,
  standardize = FALSE,
  type = c("regression", "correlation", "loading"),
  ...
)
}
\arguments{
\item{model}{CFA or SEM created by the \code{lavaan::cfa} or \code{lavaan::sem} functions.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Return standardized parameters (standardized coefficients). See \code{lavaan::standardizedsolution}.}

\item{type}{What type of links to return. Can be \code{"all"} or some of \code{c("regression", "correlation", "loading", "variance", "mean")}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data.frame of indices related to the model's parameters.
}
\description{
Format CFA/SEM objects from the (b)lavaan package (Rosseel, 2012; Merkle and Rosseel 2018).
}
\examples{
library(parameters)

# lavaan -------------------------------------
library(lavaan)

# Confirmatory Factor Analysis (CFA) ---------

structure <- " visual  =~ x1 + x2 + x3
               textual =~ x4 + x5 + x6
               speed   =~ x7 + x8 + x9 "
model <- lavaan::cfa(structure, data = HolzingerSwineford1939)
model_parameters(model)
model_parameters(model, standardize = TRUE)

# Structural Equation Model (SEM) ------------

structure <- "
  # latent variable definitions
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + a*y2 + b*y3 + c*y4
    dem65 =~ y5 + a*y6 + b*y7 + c*y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
"
model <- lavaan::sem(structure, data = PoliticalDemocracy)
model_parameters(model)
model_parameters(model, standardize = TRUE)
\donttest{
# blavaan ------------------------------------
# library(blavaan)

# model <- blavaan::bsem(structure, data=PoliticalDemocracy)
# model_parameters(model)
# model_parameters(model, standardize = TRUE)
}

}
\references{
\itemize{
  \item Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1-36.
  \item Merkle EC , Rosseel Y (2018). blavaan: Bayesian Structural Equation Models via Parameter Expansion. Journal of Statistical Software, 85(4), 1-30. http://www.jstatsoft.org/v85/i04/
}
}
