% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.lm.R
\name{model_parameters.lm}
\alias{model_parameters.lm}
\title{Linear Model Parameters}
\usage{
\method{model_parameters}{lm}(model, ci = 0.95, standardize = FALSE,
  bootstrap = FALSE, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{standardize}).}
}
\description{
Parameters of linear models.
}
\examples{
model <- lm(mpg ~ wt + cyl, data = mtcars)
model_parameters(model, standardize = TRUE)
}
\author{
\href{https://dominiquemakowski.github.io/}{Dominique Makowski}
}
