% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.merMod.R
\name{model_parameters.merMod}
\alias{model_parameters.merMod}
\title{Linear Mixed Model Parameters}
\usage{
\method{model_parameters}{merMod}(model, ci = 0.95,
  standardize = "refit", standardize_robust = FALSE,
  bootstrap = FALSE, p_method = "wald", ci_method = "wald",
  iterations = 1000, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\%).}

\item{standardize}{Add standardized parameters. Can be FALSE or a character indicating the standardization method (see \link{parameters_standardize}).}

\item{standardize_robust}{Robust standardization. See \link{parameters_standardize}.}

\item{bootstrap}{Should estimates be based on bootsrapped model? If TRUE, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply.}

\item{p_method}{Method for computing p values. See \link{p_value}.}

\item{ci_method}{Method for computing confidence intervals (CI). See \link{ci}.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{...}{Arguments passed to or from other methods (e.g., to \link[=standardize.lm]{standardize}).}
}
\description{
Parameters of linear mixed models.
}
\examples{
\dontrun{
model <- lme4::lmer(mpg ~ wt + (1 | gear), data = mtcars)
model_parameters(model, standardize = TRUE)
}

}
