% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.FactoMineR.R
\name{model_parameters.PCA}
\alias{model_parameters.PCA}
\title{FactoMineR objects Parameters}
\usage{
\method{model_parameters}{PCA}(model, ...)
}
\arguments{
\item{model}{Object of class \code{FactoMineR}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data.frame of indices related to the model's parameters.
}
\description{
Parameters of FactoMineR objects.
}
\examples{
library(FactoMineR)
model <- FactoMineR::PCA(iris[, 1:4], ncp = 2)

}
