% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R,
%   R/model_parameters.bayesian.R
\name{model_parameters}
\alias{model_parameters}
\alias{parameters}
\alias{model_parameters.brmsfit}
\title{Model Parameters}
\usage{
model_parameters(model, ...)

parameters(model, ...)

\method{model_parameters}{brmsfit}(model, centrality = "median",
  dispersion = FALSE, ci = 0.89, ci_method = "hdi", test = c("pd",
  "rope"), rope_range = "default", rope_ci = 1, bf_prior = NULL,
  diagnostic = c("ESS", "Rhat"), priors = TRUE, iterations = 1000,
  effects = "fixed", component = "all", ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}

\item{iterations}{Number of posterior draws.}

\item{effects}{Should parameters for fixed effects, random effects
or both be returned? Only applies to mixed models. May be abbreviated.}

\item{component}{Should all parameters, parameters for the
conditional model, the zero-inflated part of the model, the dispersion
term or the instrumental variables be returned? Applies to models
with zero-inflated and/or dispersion formula, or to models with instrumental
variable (so called fixed-effects regressions). May be abbreviated.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Compute and extract model parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=model_parameters.htest]{Correlations and t-tests}}
 \item{\link[=model_parameters.aov]{ANOVAs}}
 \item{\link[=model_parameters.default]{Regression models} (\code{lm}, \code{glm}, \pkg{survey}, ...)}
 \item{\link[=model_parameters.gam]{Additive models} (\code{gam}, \code{gamm}, ...)}
 \item{\link[=model_parameters.zeroinfl]{Zero-inflated models} (\code{hurdle}, \code{zeroinfl}, \code{zerocount})}
 \item{\link[=model_parameters.merMod]{Mixed models} (\pkg{lme4}, \pkg{nlme}, \pkg{glmmTMB}, ...)}
 \item{\link[=model_parameters.BFBayesFactor]{Bayesian tests} (\pkg{BayesFactor})}
 \item{\link[=model_parameters.stanreg]{Bayesian models} (\pkg{rstanarm}, \pkg{brms}, \pkg{MCMCglmm})}
 \item{\link[=model_parameters.principal]{PCA and FA} (\pkg{psych})}
 \item{\link[=model_parameters.lavaan]{CFA and SEM} (\pkg{lavaan})}
 \item{\link[=model_parameters.kmeans]{Cluster models (k-means, ...)}}
 }
}
\details{
Standardization is based on \code{\link[=effectsize::standardize_parameters]{standardize_parameters()}}.
  In case of \code{standardize = "refit"}, the data used to fit the model
  will be standardized and the model is completely refitted. In such cases,
  standard errors and confidence intervals refer to the standardized coefficient.
  For other methods of standardizing, unstandardized and standardized coefficients
  are shown, where standard errors and confidence intervals relate to the
  unstandardized coefficients.
}
\note{
The \code{\link[=print.parameters_model]{print()}} method has several arguments to tweak the output.
}
\seealso{
\code{\link[=standardize_names]{standardize_names()}} to rename
  columns into a consistent, standardized naming scheme.
}
