% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{Model Parameters}
\usage{
model_parameters(model, ...)
}
\arguments{
\item{model}{Statistical Model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute and extract model parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=model_parameters.lm]{Frequentist regressions}}
 \item{\link[=model_parameters.stanreg]{Bayesian regressions}}
 }
}
