% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.bayesian.R
\name{model_parameters.stanreg}
\alias{model_parameters.stanreg}
\title{Bayesian Models Parameters}
\usage{
\method{model_parameters}{stanreg}(model, ci = 0.9,
  standardize = FALSE, estimate = "median", test = c("pd", "rope"),
  rope_bounds = "default", rope_full = TRUE, diagnostic = TRUE,
  priors = TRUE, iterations = 1000, ...)
}
\arguments{
\item{model}{Object of class \link{lm}.}

\item{ci}{Credible Interval (CI) level. Default to 0.90 (90\%).}

\item{standardize}{Add standardized parameters. Default to FALSE as this re-fits the model and can thus take some time.}

\item{estimate}{The \href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{point-estimate(s)} to compute. Can be a character or a list with "median", "mean" or "MAP".}

\item{test}{What \href{https://easystats.github.io/bayestestR/articles/3_IndicesExistenceComparison.html}{indices of effect existence} to compute. Can be a character or a list with "p_direction", "rope" or "p_map".}

\item{rope_bounds}{\href{https://easystats.github.io/bayestestR/articles/1_IndicesDescription.html#rope}{ROPE's} lower and higher bounds. Should be a list of two values (e.g., \code{c(-0.1, 0.1)}) or \code{"default"}. If \code{"default"}, the bounds are set to \code{x +- 0.1*SD(response)}.}

\item{rope_full}{If TRUE, use the proportion of the entire posterior distribution for the equivalence test. Otherwise, use the proportion of HDI as indicated by the \code{ci} argument.}

\item{diagnostic}{Include sampling diagnostic metrics (effective sample, Rhat and MCSE). \code{Effective Sample} should be as large as possible, altough for most applications, an effective sample size greater than 1,000 is sufficient for stable estimates (Bürkner, 2017). \code{Rhat} should not be larger than 1.1.}

\item{priors}{Include priors specifications information.}

\item{iterations}{The number of bootstrap replicates. This only apply in the case of bootsrapped frequentist models.}

\item{...}{Arguments passed to or from other methods (e.g., to \code{standardize}).}
}
\description{
Parameters of Bayesian models.
}
\examples{
\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(mpg ~ wt + cyl, data = mtcars)
model_parameters(model)

library(brms)
model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
model_parameters(model)
}

}
\references{
\itemize{
 \item{\href{https://easystats.github.io/bayestestR/articles/2_IndicesEstimationComparison.html}{Comparison of Point-Estimates}}
 \item{\href{https://easystats.github.io/bayestestR/articles/3_IndicesExistenceComparison.html}{Comparison of Indices of Effect Existence}}
 }
}
