% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_wrs2.R
\name{model_parameters.t1way}
\alias{model_parameters.t1way}
\title{Parameters from \code{WRS2} objects}
\usage{
\method{model_parameters}{t1way}(model, verbose = TRUE, ...)
}
\arguments{
\item{model}{Object of class \code{t1way}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters from \code{WRS2} objects
}
\examples{
if (require("WRS2")) {
  library(WRS2)

  set.seed(123)
  model <- t1way(libido ~ dose, data = viagra)
  model_parameters(model)

  set.seed(123)
  model <- rmanova(WineTasting$Taste, WineTasting$Wine, WineTasting$Taster)
  model_parameters(model)
}
}
