% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cplm.R
\name{model_parameters.zcpglm}
\alias{model_parameters.zcpglm}
\title{Parameters from Zero-Inflated Models}
\usage{
\method{model_parameters}{zcpglm}(
  model,
  ci = 0.95,
  bootstrap = FALSE,
  iterations = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"),
  standardize = NULL,
  exponentiate = FALSE,
  robust = FALSE,
  p_adjust = NULL,
  keep = NULL,
  drop = NULL,
  parameters = keep,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model}{A model with zero-inflation component.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95\\%).}

\item{bootstrap}{Should estimates be based on bootstrapped model? If
\code{TRUE}, then arguments of \link[=model_parameters.stanreg]{Bayesian regressions} apply (see also
\code{\link[=bootstrap_parameters]{bootstrap_parameters()}}).}

\item{iterations}{The number of bootstrap replicates. This only apply in the
case of bootstrapped frequentist models.}

\item{component}{Model component for which parameters should be shown. May be
one of \code{"conditional"}, \code{"precision"} (\pkg{betareg}),
\code{"scale"} (\pkg{ordinal}), \code{"extra"} (\pkg{glmx}),
\code{"marginal"} (\pkg{mfx}), \code{"conditional"} or \code{"full"} (for
\code{MuMIn::model.avg()}) or \code{"all"}.}

\item{standardize}{The method used for standardizing the parameters. Can be
\code{"refit"}, \code{"posthoc"}, \code{"smart"}, \code{"basic"},
\code{"pseudo"} or \code{NULL} (default) for no standardization. See
'Details' in \code{\link[effectsize:standardize_parameters]{effectsize::standardize_parameters()}}.
\strong{Important:} Categorical predictors (i.e. factors) are \emph{never}
standardized by default, which may be a different behaviour compared to
other R packages or other software packages (like SPSS). If standardizing
categorical predictors is desired, either use \code{standardize="basic"}
to mimic behaviour of SPSS or packages such as \pkg{lm.beta}, or standardize
the data with \code{effectsize::standardize(force=TRUE)} before fitting
the model. Robust estimation (i.e. \code{robust=TRUE}) of standardized
parameters only works when \code{standardize="refit"}.}

\item{exponentiate}{Logical, indicating whether or not to exponentiate the
the coefficients (and related confidence intervals). This is typical for
logistic regression, or more generally speaking, for models with log
or logit links. \strong{Note:} Delta-method standard errors are also
computed (by multiplying the standard errors by the transformed
coefficients). This is to mimic behaviour of other software packages, such
as Stata, but these standard errors poorly estimate uncertainty for the
transformed coefficient. The transformed confidence interval more clearly
captures this uncertainty. For \code{compare_parameters()},
\code{exponentiate = "nongaussian"} will only exponentiate coefficients
from non-Gaussian families.}

\item{robust}{Logical, if \code{TRUE}, robust standard errors are calculated
(if possible), and confidence intervals and p-values are based on these
robust standard errors. Additional arguments like \code{vcov_estimation} or
\code{vcov_type} are passed down to other methods, see
\code{\link[=standard_error_robust]{standard_error_robust()}} for details
and \href{https://easystats.github.io/parameters/articles/model_parameters_robust.html}{this vignette}
for working examples.}

\item{p_adjust}{Character vector, if not \code{NULL}, indicates the method to
adjust p-values. See \code{\link[stats:p.adjust]{stats::p.adjust()}} for details. Further
possible adjustment methods are \code{"tukey"}, \code{"scheffe"},
\code{"sidak"} and \code{"none"} to explicitly disable adjustment for
\code{emmGrid} objects (from \pkg{emmeans}).}

\item{keep}{Character containing a regular expression pattern
that describes the parameters that should be included in the returned data
frame (for \code{keep}), resp. parameters to exclude (\code{drop}). \code{keep}
may also be a named list of regular expressions. All non-matching parameters
will be removed from the output. If \code{keep} is a character vector, every
parameter name in the \emph{"Parameter"} column that matches the regular expression
in \code{parameters} will be selected from the returned data frame (and vice
versa, all parameter names matching \code{drop} will be excluded). Furthermore,
if \code{keep} has more than one element, these will be merged with an
\code{OR} operator into a regular expression pattern like this:
\code{"(one|two|three)"}. If \code{keep} is a named list of regular expression
patterns, the names of the list-element should equal the column name where
selection should be applied. This is useful for model objects where
\code{model_parameters()} returns multiple columns with parameter components,
like in \code{\link[=model_parameters.lavaan]{model_parameters.lavaan()}}. Note that the regular expression
pattern should match the parameter names as they are stored in the returned
data frame, which can be different from how they are printed. Inspect the
\verb{$Parameter} column of the parameters table to get the exact parameter
names.}

\item{drop}{Character containing a regular expression pattern
that describes the parameters that should be included in the returned data
frame (for \code{keep}), resp. parameters to exclude (\code{drop}). \code{keep}
may also be a named list of regular expressions. All non-matching parameters
will be removed from the output. If \code{keep} is a character vector, every
parameter name in the \emph{"Parameter"} column that matches the regular expression
in \code{parameters} will be selected from the returned data frame (and vice
versa, all parameter names matching \code{drop} will be excluded). Furthermore,
if \code{keep} has more than one element, these will be merged with an
\code{OR} operator into a regular expression pattern like this:
\code{"(one|two|three)"}. If \code{keep} is a named list of regular expression
patterns, the names of the list-element should equal the column name where
selection should be applied. This is useful for model objects where
\code{model_parameters()} returns multiple columns with parameter components,
like in \code{\link[=model_parameters.lavaan]{model_parameters.lavaan()}}. Note that the regular expression
pattern should match the parameter names as they are stored in the returned
data frame, which can be different from how they are printed. Inspect the
\verb{$Parameter} column of the parameters table to get the exact parameter
names.}

\item{parameters}{Deprecated, alias for \code{keep}.}

\item{verbose}{Toggle warnings and messages.}

\item{...}{Arguments passed to or from other methods. For instance, when
\code{bootstrap = TRUE}, arguments like \code{type} or \code{parallel} are
passed down to \code{bootstrap_model()}, and arguments like \code{ci_method}
are passed down to \code{\link[bayestestR:describe_posterior]{bayestestR::describe_posterior()}}.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
Parameters from zero-inflated models (from packages like \pkg{pscl},
\pkg{cplm} or \pkg{countreg}).
}
\examples{
library(parameters)
if (require("pscl")) {
  data("bioChemists")
  model <- zeroinfl(art ~ fem + mar + kid5 + ment | kid5 + phd, data = bioChemists)
  model_parameters(model)
}
}
\seealso{
\code{\link[insight:standardize_names]{insight::standardize_names()}} to rename
columns into a consistent, standardized naming scheme.
}
