% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulate.R
\name{model_simulate}
\alias{model_simulate}
\alias{model_simulate.glmmTMB}
\alias{model_simulate.zeroinfl}
\title{Simulated draws from model coefficients}
\usage{
model_simulate(model, n_sims = 1000, ...)

\method{model_simulate}{glmmTMB}(model, n_sims = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"), ...)

\method{model_simulate}{zeroinfl}(model, n_sims = 1000,
  component = c("all", "conditional", "zi", "zero_inflated"), ...)
}
\arguments{
\item{model}{Statistical model.}

\item{n_sims}{The number of simulation draws to create.}

\item{...}{Arguments passed to or from other methods.}

\item{component}{Should all parameters, parameters for the conditional model,
or for the zero-inflated part of the model be returned? Applies to models
with zero-inflated component. \code{component} may be one of \code{"conditional"},
\code{"zi"}, \code{"zero-inflated"} or \code{"all"} (default). May be abbreviated.}
}
\value{
A data.frame.
}
\description{
Simulate draws a statistical model n times to return a data.frame of estimates.
}
\details{
\code{model_simulate()} is a computationally faster alternative
  to \code{model_bootstrap()}. Simulated draws for coefficients are based
  on a multivariate normal distribution (\code{MASS::mvrnorm()} with mean
  \code{coef(model)} and variance \code{vcov(model)}.
}
\examples{
library(parameters)

model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
head(model_simulate(model))
}
\seealso{
\code{\link[=parameters_simulate]{parameters_simulate()}},
\code{\link[=model_bootstrap]{model_bootstrap()}},
\code{\link[=parameters_bootstrap]{parameters_bootstrap()}}
}
