% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_factors.R
\name{n_factors}
\alias{n_factors}
\title{Number of Components/Factors to Retain}
\usage{
n_factors(x, type = "FA", rotation = "varimax",
  algorithm = "default", package = "all", safe = TRUE, ...)
}
\arguments{
\item{x}{A dataframe.}

\item{type}{Can be "FA" or "PCA", depending on what you want to do.}

\item{rotation}{Only used for VSS. The rotation to apply. Can be "none" "varimax", "quartimax", "bentlerT", "equamax", "varimin", "geominT" and "bifactor" for orthogonal rotations, and "promax", "oblimin", "simplimax", "bentlerQ, "geominQ" and "biquartimin" and "cluster" for oblique transformations.}

\item{algorithm}{Factoring method used by VSS. Can be "pa" for Principal Axis Factor Analysis, "minres" for minimum residual (OLS) factoring, "mle" for Maximum Likelihood FA and "pc" for Principal Components. "default" will select "minres" if \code{type} is "FA" and "pc" if \code{type} is "PCA".}

\item{package}{Can be "all" or a vector containing "nFactors", "EGAnet" and "psych". These are the packages from which methods are used.}

\item{safe}{If TRUE, will run all the procedures in try blocks, and will only return those that work and silently skip the ones that may fail.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function runs many existing procedures for determining how many factors to retain for your factor analysis (FA) or dimension reduction (PCA).
}
\examples{
library(parameters)

n_factors(mtcars, type = "PCA", package = "nFactors")

result <- n_factors(mtcars[, 1:5], type = "FA", package = "psych")
as.numeric(result)
summary(result)

\dontrun{
  n_factors(mtcars, type = "PCA")
  n_factors(mtcars, type = "FA", algorithm = "mle")
}

}
