% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_to_probs.R
\name{odds_to_probs}
\alias{odds_to_probs}
\title{Convert (log)odds to probabilies.}
\usage{
odds_to_probs(odds, select = NULL, exclude = NULL, log = FALSE)
}
\arguments{
\item{odds}{Odds values in vector or dataframe.}

\item{select}{Character or list of of column names to be
transformed.}

\item{exclude}{Character or list of column names to be excluded
from transformation.}

\item{log}{Are these Log odds (such as in logistic models)?}
}
\description{
Convert (log)odds to probabilies.
}
\examples{
odds_to_probs(-1.45)

}
