% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\title{Computation of p-values}
\usage{
p_value(model, ...)
}
\arguments{
\item{model}{Statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Computes p-values related to parameters.
}
