% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\alias{p_value.lmerMod}
\alias{p_value.glmmTMB}
\title{p-values}
\usage{
p_value(model, ...)

\method{p_value}{lmerMod}(model, method = "wald", ...)

\method{p_value}{glmmTMB}(model, component = c("all", "conditional",
  "zi", "zero_inflated"), ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}

\item{method}{For mixed models, can be \link[=p_value_wald]{"wald"} (default) or \link[=p_value_kenward]{"kenward"}.}

\item{component}{Should all parameters, parameters for the conditional model,
or for the zero-inflated part of the model be returned? Applies to models
with zero-inflated component. \code{component} may be one of \code{"conditional"},
\code{"zi"}, \code{"zero-inflated"} or \code{"all"} (default). May be abbreviated.}
}
\value{
The p-values.
}
\description{
This function attempts to return, or compute, p-values of a model's parameters. The nature of the p-values is different depending on the model:
\itemize{
\item Mixed models (lme4): TO BE IMPROVED.
}
}
\examples{
model <- lme4::lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
p_value(model)
}
