% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{p_value}
\alias{p_value}
\title{Compute p-values}
\usage{
p_value(model, method = "wald", ...)
}
\arguments{
\item{model}{A statistical model.}

\item{method}{For mixed models, can be \link{p_value_wald}{"wald"} (default) or \link{p_value_kenward}{"kenward"}.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function attempts to return, or compute, p-values of a model's parameters. The nature of the p-values is different depending on the model:
\itemize{
\item Mixed models (lme4): TO BE IMPROVED.
}
}
\examples{
\dontrun{
model <- lme4::lmer(Petal.Length ~ Sepal.Length + (1|Species), data=iris)
p_value(model)
}
}
