% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value_wald.R
\name{p_value_wald}
\alias{p_value_wald}
\title{p-values using Wald-test approximation}
\usage{
p_value_wald(model)
}
\arguments{
\item{model}{A statistical model.}
}
\description{
The Wald-test approximation treats t-values as Wald z. Since the t distribution converges to the z distribution as degrees of freedom increase, this is like assuming infinite degrees of freedom. While this is unambiguously anti-conservative, this approximation appears as reasonable for reasonable sample sizes (Barr et al., 2013). That is, if we take the p-value to measure the probability of a false positive, this approximation produces a higher false positive rate than the nominal 5% at p = 0.05.
}
\examples{
\dontrun{
library(lme4)
model <- lmer(Petal.Length ~ Sepal.Length + (1 | Species), data = iris)
p_value_wald(model)
}
}
