\name{Parameters-class}
\docType{class}

\alias{class:Parameters}
\alias{Parameters-class}

\alias{getAll,Parameters-method}
\alias{value,Parameters-method}
\alias{value<-,Parameters-method}

\title{Parameters}

\description{
Implements a simple YAML-based approach to possibly nested lists of
parameter (name-value) assignemnts.
}


\section{Methods}{
  \describe{
    \item{}{
      \code{value(x, category, name)}: extract matrices by metadata.
     }
   \item{}{
      \code{getAll(x)}: returns nested list
    }
   }
 }

\author{Paul Shannon}

\examples{

}
\keyword{methods}
\keyword{classes}
