\name{Parameters-package}
\Rdversion{1.1}
\alias{Parameters-package}

\docType{package}


\title{
  Parameters: general purpose YAML-based parameter access
}
\description{

  A simple utility class for retrieving parameter values from a YAML
  file, modifying and querying these values as needed.  It will be most
  useful to package builders whose running code uses a large number
  of run-time parameters and 'tuning' variables.

}
\details{

  Instructions for use are in the vignette, available with
  \code{browseVignettes(package="Parameters")}

}
\author{
Author: Paul Shannon

Maintainer: Biocore Team c/o BioC user list <bioconductor@stat.math.ethz.ch>
}

\keyword{ package}

