% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_bootstrap.R
\name{parameters_bootstrap}
\alias{parameters_bootstrap}
\title{Parameters bootstrapping}
\usage{
parameters_bootstrap(model, iterations = 1000, centrality = "median",
  ci = 0.95, ci_method = "quantile", test = "p-value", ...)
}
\arguments{
\item{model}{Statistical model.}

\item{iterations}{The number of bootstrap replicates.}

\item{centrality}{The point-estimates (centrality indices) to compute.  Character (vector) or list with one or more of these options: \code{"median"}, \code{"mean"}, \code{"MAP"} or \code{"all"}.}

\item{ci}{Value or vector of probability of the CI (between 0 and 1)
to be estimated. Default to \code{.89} (89\%) for Bayesian models and \code{.95} (95\%) for frequentist models.}

\item{ci_method}{The type of index used for Credible Interval. Can be
\code{"hdi"} (default, see \code{\link{hdi}}) or \code{"quantile"}
(see \code{\link{ci}}).}

\item{test}{The indices of effect existence to compute. Character (vector) or
list with one or more of these options: \code{"p_direction"} (or \code{"pd"}),
\code{"rope"}, \code{"p_map"}, \code{"equivalence_test"} (or \code{"equitest"}),
\code{"bayesfactor"} (or \code{"bf"}) or \code{"all"} to compute all tests.#'
For each "test", the corresponding \pkg{bayestestR} function is called
(e.g. \code{\link{rope}} or \code{\link{p_direction}}) and its results
included in the summary output.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute bootstrapped parameters and their related indices such as Confidence Intervals (CI) and p-values.
}
\examples{
library(parameters)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
parameters_bootstrap(model)
}
\references{
Davison, A. C., & Hinkley, D. V. (1997). Bootstrap methods and their application (Vol. 1). Cambridge university press.
}
