% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_groupspecific.R
\name{parameters_groupspecific}
\alias{parameters_groupspecific}
\title{Extract group-specific parameters of mixed models}
\usage{
parameters_groupspecific(model, indices = c("Coefficient", "SE"), ...)
}
\arguments{
\item{model}{A mixed model.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Extract parameters of each individual group in the context of mixed models.
}
\examples{
library(parameters)
if (require("lme4")) {
  data <- lme4::sleepstudy

  # Random intercept
  model <- lmer(Reaction ~ Days + (1 | Subject), data = data)
  random <- parameters_groupspecific(model)
  head(random)
  summary(random)

  # Random slope and intercept
  model <- lmer(Reaction ~ Days + (1 + Days  | Subject), data = data)
  random <- parameters_groupspecific(model)
  head(random)

  # Multiple random groups
  model <- lmer(mpg ~ wt + (1 | gear) + (1 | carb), data = mtcars)
  random <- parameters_groupspecific(model)
  head(random)

  # Nested random factors
  set.seed(33)
  data$grp <- sample(letters[1:5], size = 180, replace = TRUE)
  data$subgrp <- NA
  for (i in letters[1:5]) {
    filter_group <- data$grp == i
    data$subgrp[filter_group] <- sample(LETTERS, size = sum(filter_group), replace = TRUE)
  }
  model <- lmer(Reaction ~ Days + (1 | grp / subgrp) + (1 | Subject), data = data)
  # parameters_groupspecific(model)

}
}
