% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_table.R
\name{parameters_table}
\alias{parameters_table}
\title{Parameters Table Formatting}
\usage{
parameters_table(x, ...)
}
\arguments{
\item{x}{A dataframe of model's parameters.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Parameters Table Formatting
}
\examples{
library(parameters)

model_parameters(lm(Sepal.Length ~ Species * Sepal.Width, data = iris))

}
