% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_type.R
\name{parameters_type}
\alias{parameters_type}
\title{Type of Model Parameters}
\usage{
parameters_type(model, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data.frame.
}
\description{
Type of Model Parameters
}
\examples{
library(parameters)
model <- lm(Sepal.Length ~ Species * Sepal.Width * Petal.Length, data = iris)

parameters_type(model)

}
