% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_parameters.R
\name{pool_parameters}
\alias{pool_parameters}
\title{Pool Model Parameters}
\usage{
pool_parameters(x, exponentiate = FALSE, ...)
}
\arguments{
\item{x}{A list of \code{parameters_model} objects, as returned by
\code{\link{model_parameters}}, or a list of model-objects that is
supported by \code{model_parameters()}.}

\item{exponentiate}{Logical, indicating whether or not to exponentiate the the coefficients (and related confidence intervals). This is typical for, say, logistic regressions, or more generally speaking: for models with log or logit link. \strong{Note:} standard errors are also transformed (by multiplying the standard errors with the exponentiated coefficients), to mimic behaviour of other software packages, such as Stata.}

\item{...}{Currently not used.}
}
\value{
A data frame of indices related to the model's parameters.
}
\description{
This function "pools" (i.e. combines) model parameters in a similar fashion as \code{mice::pool()}. However, this function pools parameters from \code{parameters_model} objects, as returned by \code{\link{model_parameters}}.
}
\details{
Averaging of parameters follows Rubin's rules (\cite{Rubin, 1987, p. 76}).
  The pooled degrees of freedom is based on the Barnard-Rubin adjustment for
  small samples (\cite{Barnard and Rubin, 1999}).
}
\note{
This function is still experimental. Results should be reliable,
  however, this function may not yet work with all types of models properly.
  In particular models with identical names for coefficients (for instance,
  models with zero-inflation, where predictors appear in the count and
  zero-inflated part) may fail, since the coefficient table is grouped by
  coefficient names for pooling. In such cases, coefficients of count and
  zero-inflated model parts would be combined.
}
\examples{
# example for multiple imputed datasets
if (require("mice")) {
  data("nhanes2")
  imp <- mice(nhanes2)
  models <- lapply(1:5, function(i) {
    lm(bmi ~ age + hyp + chl, data = complete(imp, action = i))
  })
  pool_parameters(models)

  # should be identical to:
  m <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
  summary(pool(m))
}
}
\references{
Barnard, J. and Rubin, D.B. (1999). Small sample degrees of freedom with multiple imputation. Biometrika, 86, 948-955.
Rubin, D.B. (1987). Multiple Imputation for Nonresponse in Surveys. New York: John Wiley and Sons.
}
