% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_components.R
\name{principal_components}
\alias{principal_components}
\alias{PCA}
\title{Principal Components Analysis (PCA)}
\usage{
principal_components(x, n = NULL, threshold = NULL,
  standardize = TRUE, ...)

PCA(x, n = NULL, threshold = NULL, standardize = TRUE, ...)
}
\arguments{
\item{x}{A dataframe.}

\item{n}{Number of components to extract. If \code{n = NULL}, the number of components is selected through \code{\link{n_factors}}.}

\item{threshold}{A value between 0 and 1 indicating which (absolute) values from the loadings should be removed.}

\item{standardize}{A logical value indicating whether the variables should be standardized (centred and scaled) to have unit variance before the analysis takes place (in general, such scaling is advisable).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function performs a principal component analysis (PCA) and returns the loadings (of the unrotated matrix) as dataframe.
}
\examples{
library(parameters)

principal_components(mtcars)
principal_components(mtcars, n = "all", threshold = 0.2)

}
