% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence_test.R, R/methods_bayestestR.R,
%   R/n_parameters.R, R/p_direction.R, R/p_significance.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{equivalence_test}
\alias{ci}
\alias{n_parameters}
\alias{p_direction}
\alias{p_significance}
\alias{standardize_names}
\alias{supported_models}
\alias{print_html}
\alias{print_md}
\alias{display}
\alias{describe_distribution}
\alias{demean}
\alias{rescale_weights}
\alias{visualisation_recipe}
\alias{kurtosis}
\alias{skewness}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}, \code{\link[bayestestR]{p_direction}}, \code{\link[bayestestR]{p_significance}}}

  \item{datawizard}{\code{\link[datawizard]{demean}}, \code{\link[datawizard]{describe_distribution}}, \code{\link[datawizard:skewness]{kurtosis}}, \code{\link[datawizard]{rescale_weights}}, \code{\link[datawizard]{skewness}}, \code{\link[datawizard]{visualisation_recipe}}}

  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight]{n_parameters}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}, \code{\link[insight]{standardize_names}}, \code{\link[insight:is_model_supported]{supported_models}}}
}}

