% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_factorstructure.R, R/display.R,
%   R/equivalence_test.R, R/methods_bayestestR.R, R/n_parameters.R,
%   R/print_html.R, R/print_md.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{check_sphericity}
\alias{display}
\alias{equivalence_test}
\alias{ci}
\alias{n_parameters}
\alias{print_html}
\alias{print_md}
\alias{standardize_names}
\alias{supported_models}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}}

  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight]{n_parameters}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}, \code{\link[insight]{standardize_names}}, \code{\link[insight:is_model_supported]{supported_models}}}

  \item{performance}{\code{\link[performance]{check_sphericity}}}
}}

