% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/display.R, R/equivalence_test.R,
%   R/methods_bayestestR.R, R/print_md.R, R/re-exports.R, R/standardize_names.R
=======
% Please edit documentation in R/ci.R, R/display.R, R/equivalence_test.R,
%   R/print_html.R, R/print_md.R, R/re-exports.R, R/standardize_names.R
>>>>>>> upstream/master
\docType{import}
\name{reexports}
\alias{reexports}
\alias{display}
\alias{equivalence_test}
<<<<<<< HEAD
\alias{ci}
=======
\alias{print_html}
>>>>>>> upstream/master
\alias{print_md}
\alias{format_bf}
\alias{format_pd}
\alias{parameters_table}
\alias{format_p}
\alias{standardize_names}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}}

  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight]{format_bf}}, \code{\link[insight]{format_p}}, \code{\link[insight]{format_pd}}, \code{\link[insight]{parameters_table}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}, \code{\link[insight]{standardize_names}}}
}}

