% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R, R/equivalence_test.R,
%   R/methods_bayestestR.R, R/n_parameters.R, R/print_html.R, R/print_md.R,
%   R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{display}
\alias{equivalence_test}
\alias{ci}
\alias{n_parameters}
\alias{print_html}
\alias{print_md}
\alias{standardize_names}
\alias{supported_models}
\alias{demean}
\alias{data_to_numeric}
\alias{skewness}
\alias{kurtosis}
\alias{describe_distribution}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}}

  \item{datawizard}{\code{\link[datawizard:convert_data_to_numeric]{data_to_numeric}}, \code{\link[datawizard]{demean}}, \code{\link[datawizard]{describe_distribution}}, \code{\link[datawizard:skewness]{kurtosis}}, \code{\link[datawizard]{skewness}}}

  \item{insight}{\code{\link[insight]{display}}, \code{\link[insight]{n_parameters}}, \code{\link[insight:display]{print_html}}, \code{\link[insight:display]{print_md}}, \code{\link[insight]{standardize_names}}, \code{\link[insight:is_model_supported]{supported_models}}}
}}

