% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R, R/equivalence_test.R, R/re-exports.R,
%   R/standardize_names.R, R/to_table.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ci}
\alias{equivalence_test}
\alias{format_number}
\alias{format_p}
\alias{format_rope}
\alias{format_bf}
\alias{format_pd}
\alias{standardize_names}
\alias{to_table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bayestestR}{\code{\link[bayestestR]{ci}}, \code{\link[bayestestR]{equivalence_test}}}

  \item{insight}{\code{\link[insight]{format_bf}}, \code{\link[insight]{format_number}}, \code{\link[insight]{format_p}}, \code{\link[insight]{format_pd}}, \code{\link[insight]{format_rope}}, \code{\link[insight]{standardize_names}}, \code{\link[insight]{to_table}}}
}}

