% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness_kurtosis.R
\name{skewness}
\alias{skewness}
\alias{kurtosis}
\title{Compute Skewness and Kurtosis}
\usage{
skewness(x, na.rm = TRUE)

kurtosis(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector or data.frame.}

\item{na.rm}{Remove NaNs.}
}
\description{
Compute Skewness and Kurtosis
}
\examples{
skewness(rnorm(1000))
kurtosis(rnorm(1000))
}
\references{
All credits go to the \code{moments} package.
}
