% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Extract standard errors}
\usage{
standard_error(model, ...)
}
\arguments{
\item{model}{A model.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame.
}
\description{
This function attempts to return standard errors of model parameters.
}
\examples{
model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
standard_error(model)
}
