% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.data.frame.R
\name{standardize.data.frame}
\alias{standardize.data.frame}
\title{Data Standardization}
\usage{
\method{standardize}{data.frame}(x, robust = FALSE, select = NULL,
  exclude = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{select}{For a data.frame, character or list of characters of column names to be
standardized.}

\item{exclude}{For a data.frame, character or list of characters of column names to be excluded from standardization.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Standardize (scale and reduce, Z-score) the data so that the values are expressed in terms of standard deviation (i.e., mean = 0, SD = 1) or Median Absolute Deviance (\code{robust = TRUE}; median = 0, MAD = 1).
}
\examples{
summary(standardize(iris))
}
