% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.models.R
\name{standardize.lm}
\alias{standardize.lm}
\title{Model Standardization}
\usage{
\method{standardize}{lm}(x, method = "mean", ...)
}
\arguments{
\item{x}{Object.}

\item{method}{The method of standardization. See 'Details'.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Refit the model after standardizing the data.
}
\examples{
model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
coef(standardize(model))

}
