% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.data.frame.R
\name{standardize.numeric}
\alias{standardize.numeric}
\title{Standardization}
\usage{
\method{standardize}{numeric}(x, robust = FALSE, method = "default",
  verbose = TRUE, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{Logical, if \code{TRUE}, centering is done by substracting the
median from the variables and divide it by the median absolute deviation
(MAD). If \code{FALSE}, variables are standardized by substracting the
mean and divide it by the standard deviation (SD).
the}

\item{method}{The method of standardization. See 'Details'.}

\item{verbose}{Toggle warnings on or off.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Performs a standardization of data or parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=standardize.data.frame]{Dataframes}}
 \item{\link[=standardize.lm]{Regression models}}
 }
}
\details{
Methods:
\itemize{
 \item \strong{default}: Depending on \code{robust}, variables are divided by SD or MAD.
 \item \strong{2sd}: Depending on \code{robust}, variables are divided by two times the SD or MAD.
}
}
