% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardization}
\usage{
standardize(x, robust = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Performs a standardization of data or parameters. See the documentation for your object's class:
\itemize{
 \item{\link[=standardize.data.frame]{data.frame}}
 \item{\link[=standardize.lm]{lm}}
 }
}
