% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_info.R
\name{standardize_info}
\alias{standardize_info}
\alias{standardise_info}
\title{Get Standardization Information}
\usage{
standardize_info(
  model,
  robust = FALSE,
  two_sd = FALSE,
  include_pseudo = FALSE,
  ...
)
}
\arguments{
\item{model}{A statistical model.}

\item{include_pseudo}{(For (G)LMMs) Should Pseudo-standardized information be
included?}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A data frame with information on each parameter (see
\code{\link[=parameters_type]{parameters_type()}}), and various standardization coefficients
for the post-hoc methods (see \code{\link[=standardize_parameters]{standardize_parameters()}}) for the predictor
and the response.
}
\description{
This function extracts information, such as the deviations (SD or MAD) from
parent variables, that are necessary for post-hoc standardization of
parameters. This function gives a window on how standardized are obtained,
i.e., by what they are divided. The "basic" method of standardization uses.
}
\examples{
model <- lm(mpg ~ ., data = mtcars)
standardize_info(model)
standardize_info(model, robust = TRUE)
standardize_info(model, two_sd = TRUE)
}
\seealso{
Other standardize: 
\code{\link{standardize_parameters}()}
}
\concept{standardize}
