% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_parameters.R
\name{standardize_parameters}
\alias{standardize_parameters}
\title{Get Standardized Model Parameters}
\usage{
standardize_parameters(model, robust = FALSE, method = "refit",
  verbose = TRUE, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{robust}{Logical, if \code{TRUE}, centering is done by substracting the
median from the variables and divide it by the median absolute deviation
(MAD). If \code{FALSE}, variables are standardized by substracting the
mean and divide it by the standard deviation (SD).
the}

\item{method}{The method used for standardizing the parameters. Can be \code{"refit"} (default).}

\item{verbose}{Toggle warnings on or off.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute standardized model parameters (coefs).
}
\details{
\strong{Methods:}
\itemize{
 \item \strong{refit}: This method is based on a complete model re-fit using the standardized data. It is the most accurate, especially for parameters related to interactions, but it is also the most computationnaly costly.
 \item \strong{2sd}: Same as \code{method = "refit"}, however, standardization is done by dividing by two times the SD or MAD (depending on \code{robust}).
 \item \strong{full}: Post-hoc standardization of the model parmaters.
 \item \strong{classic}: Similar to \code{method = "full"}, but factors are treated differently. See below.
}
\cr \cr
\strong{Standardization of factors}
}
\examples{
data(iris)

model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
standardize_parameters(model, method = "refit")
standardize_parameters(model, method = "refit", robust = TRUE)
standardize_parameters(model, method = "2sd")
standardize_parameters(model, method = "2sd", robust = TRUE)
standardize_parameters(model, method = "full")
standardize_parameters(model, method = "full", robust = TRUE)

iris$binary <- ifelse(iris$Sepal.Width > 3, 1, 0)
model <- glm(binary ~ Species * Sepal.Length, data = iris, family = "binomial")
standardize_parameters(model, method = "refit")
standardize_parameters(model, method = "refit", robust = TRUE)
standardize_parameters(model, method = "full")
standardize_parameters(model, method = "full", robust = TRUE)

}
