% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_parameters.R
\name{standardize_parameters}
\alias{standardize_parameters}
\title{Get Standardized Model Parameters}
\usage{
standardize_parameters(model, method = "refit", robust = FALSE, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{method}{The method used for standardizing the parameters. Can be "refit" (default).}

\item{robust}{If TRUE, the standardization will be based on \link{median} and \link{mad} instead of \link{mean} and \link{sd} (default).}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Compute standardized model parameters (coefs).
}
\details{
Methods:
\itemize{
 \item \emph{refit}: This method is based on a complete model re-fit using the standardized data. It is the most accurate, especially for parameters related to interactions, but it is also the most computationnaly costly.
 }
}
\examples{
model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
standardize_parameters(model, method="refit")
standardize_parameters(model, method="refit", robust=TRUE)
standardize_parameters(model, method="full", robust=FALSE)
standardize_parameters(model, method="full", robust=TRUE)


}
