% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{methods}
\name{value}
\alias{value}
\alias{value,parameter-method}
\alias{value<-}
\alias{value<-,parameter-method}
\title{value}
\usage{
value(object)

\S4method{value}{parameter}(object)

value(object) <- value

\S4method{value}{parameter}(object) <- value
}
\arguments{
\item{object}{A \code{parameter} object}

\item{value}{A numeric vector of length one. Must be fall between the parameter
objects upper and lower bounds.}
}
\description{
Extract or replact the value of parameter object
}
\section{Methods (by class)}{
\itemize{
\item \code{parameter}: Returns the value held the in the \code{value} slot as a
named numeric vector.

\item \code{parameter}: Replaces the value held the in the \code{value} slot
}}
\examples{
p <- new("parameter", value = .5, name = "binomial_p",
         upper_bound = 1, lower_bound = 0)
value(p)
value(p) <- .9
value(p)
}

